% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquifer_utilities.R
\name{add.well<-}
\alias{add.well<-}
\title{add.well<-}
\usage{
add.well(x) <- value
}
\arguments{
\item{x}{An aquifer object}

\item{value}{A list with the values of the well discharges, x coordinates 
of the wells and the y coordinates of the wells}
}
\value{
This function adds a well to an aquifer
}
\description{
Function to add wells to an aquifer object
}
\seealso{
Other aquifer functions: \code{\link{aquifer}},
  \code{\link{calculate_drawdown}},
  \code{\link{print.aquifer}},
  \code{\link{summary.aquifer}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{aquifer functions}
