% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest_data.R
\docType{data}
\name{agarwal_recovery}
\alias{agarwal_recovery}
\title{estimation of a recovery test after a constant rate or variable rate test.}
\format{A data frame with 15 rows and 2 variables
\describe{
\item{t}{Time in seconds}
\item{s}{Drawdown in m}
}}
\source{
Batu, V., Aquifer Hydraulics: A Comprehensive Guide to Hydrogeologic
Data Analysis, John Wiley, New York, 1998. Example 4-12, Pages 183-186
}
\usage{
agarwal_recovery
}
\description{
The information of this recovery test is the following:
\itemize{
\item Duration of the pumping test (in s): t <- 14400
\item Pumping rate (in m3/s): Q <- 0.02893519
\item Distance to the pumping well (in m): r <- 60
}
}
\keyword{datasets}
