% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agarwal_recovery_utilities.R
\name{agarwal_recovery_calculate_parameters}
\alias{agarwal_recovery_calculate_parameters}
\title{agarwal_recovery_calculate_parameters}
\usage{
agarwal_recovery_calculate_parameters(ptest, par, hydraulic = TRUE)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{par}{A list with the values of slope and intercept (a and t0) of the straight
line fitted to the drawdown data using the Cooper-Jacob approach}

\item{hydraulic}{Logical flag to indicate if hydraulic parameters are calculated. If False,
the the statistcal parameter (a and t0) are calculated.}
}
\value{
A list with the hydraulic parameters:
\itemize{
\item Tr: Transmissivity
\item Ss: Storage coefficient
\item radius_influence
}
}
\description{
Wrapper to function theis_calculate_parameters (The recovery test is analyzed using the Theis solution)
}
\examples{
data(agarwal_recovery)
pump_history <- 14400
agarwal_corr <- recovery_equivalent_time(agarwal_recovery$t, agarwal_recovery$s,
                                pump_history)
ptest.agarwal <- pumping_test("Well1", Q = 0.02893519, r = 60, t = agarwal_corr$ta,
s = agarwal_corr$sa)
sol0 <- agarwal_recovery_solution_initial(ptest.agarwal)
hydr.par <- agarwal_recovery_calculate_parameters(ptest.agarwal, sol0, TRUE)
}
\references{
Agarwal, R. (1980). A New Method To Account For Producing Time Effects When Drawdown
Type Curves Are Used To Analyze Pressure Buildup And Other Test Data. SPE Paper 9289.
}
\seealso{
Other agarwal_recovery functions: \code{\link{agarwal_recovery_WF_LT}},
  \code{\link{agarwal_recovery_solution_dlogt}},
  \code{\link{agarwal_recovery_solution_initial}},
  \code{\link{agarwal_recovery_solution}},
  \code{\link{agarwal_recovery_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{agarwal_recovery functions}
