% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agarwal_recovery_utilities.R
\name{agarwal_recovery_solution}
\alias{agarwal_recovery_solution}
\title{agarwal_recovery_solution}
\usage{
agarwal_recovery_solution(ptest, a, t0, t)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{a}{Slope of the straight line fitted to the drawdown data using the
Cooper-Jacob approach}

\item{t0}{Intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach}

\item{t}{Numeric vector with time}
}
\value{
A numeric vector with the calculate drawdown
}
\description{
Wrapper to function theis_solution (The recovery test is analyzed using the Theis solution)
}
\examples{
data(agarwal_recovery)
pump_history <- 14400
agarwal_corr <- recovery_equivalent_time(agarwal_recovery$t, agarwal_recovery$s,
                                pump_history)
ptest.agarwal <- pumping_test("Well1", Q = 0.02893519, r = 60, t = agarwal_corr$ta,
s = agarwal_corr$sa)
sol0 <- agarwal_recovery_solution_initial(ptest.agarwal)
scal <- agarwal_recovery_solution(ptest.agarwal, sol0$a, sol0$t, ptest.agarwal$t)
print(scal)
}
\references{
Agarwal, R. (1980). A New Method To Account For Producing Time Effects When Drawdown
Type Curves Are Used To Analyze Pressure Buildup And Other Test Data. SPE Paper 9289.
}
\seealso{
Other agarwal_recovery functions: \code{\link{agarwal_recovery_WF_LT}},
  \code{\link{agarwal_recovery_calculate_parameters}},
  \code{\link{agarwal_recovery_solution_dlogt}},
  \code{\link{agarwal_recovery_solution_initial}},
  \code{\link{agarwal_recovery_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{agarwal_recovery functions}
