% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agarwal_recovery_utilities.R
\name{agarwal_recovery_solution_dlogt}
\alias{agarwal_recovery_solution_dlogt}
\title{agarwal_recovery_solution_dlogt}
\usage{
agarwal_recovery_solution_dlogt(ptest, a, t0, t)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{a}{Slope of the straight line fitted to the drawdown data using the
Cooper-Jacob approach}

\item{t0}{Intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach}

\item{t}{Numeric vector with the values of time}
}
\value{
A list with
\itemize{
\item x: Numeric vector with the time values
\item y: Numeric vector with the derivative values
}
}
\description{
Wrapper to function theis_solution_dlogt (The recovery test is analyzed using the Theis solution)
}
\examples{
data(agarwal_recovery)
}
\references{
Agarwal, R. (1980). A New Method To Account For Producing Time Effects When Drawdown
Type Curves Are Used To Analyze Pressure Buildup And Other Test Data. SPE Paper 9289.
}
\seealso{
Other agarwal_recovery functions: \code{\link{agarwal_recovery_WF_LT}},
  \code{\link{agarwal_recovery_calculate_parameters}},
  \code{\link{agarwal_recovery_solution_initial}},
  \code{\link{agarwal_recovery_solution}},
  \code{\link{agarwal_recovery_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{agarwal_recovery functions}
