% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agarwal_recovery_utilities.R
\name{agarwal_recovery_well_function}
\alias{agarwal_recovery_well_function}
\title{agarwal_recovery_well_function}
\usage{
agarwal_recovery_well_function(td, par = NULL)
}
\arguments{
\item{td}{Numeric vector with dimensionless time}

\item{par}{Additional parameters}
}
\value{
A numeric vector with the dimensionless drawdown
}
\description{
Wrapper to function theis_well_function (The recovery test is analyzed using the Theis solution)
}
\examples{
data(agarwal_recovery)
}
\references{
Agarwal, R. (1980). A New Method To Account For Producing Time Effects When Drawdown
Type Curves Are Used To Analyze Pressure Buildup And Other Test Data. SPE Paper 9289.
}
\seealso{
Other agarwal_recovery functions: \code{\link{agarwal_recovery_WF_LT}},
  \code{\link{agarwal_recovery_calculate_parameters}},
  \code{\link{agarwal_recovery_solution_dlogt}},
  \code{\link{agarwal_recovery_solution_initial}},
  \code{\link{agarwal_recovery_solution}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{agarwal_recovery functions}
