% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agarwal_recovery_utilities.R
\name{agarwal_recovery_WF_LT}
\alias{agarwal_recovery_WF_LT}
\title{agarwal_recovery_WF_LT}
\usage{
agarwal_recovery_WF_LT(p, ...)
}
\arguments{
\item{p}{Laplace Transform parameter}

\item{...}{Additional parameters}
}
\value{
A numeric vector with the calculate drawdown in the Laplace domain
}
\description{
Wrapper to function theis_WF_LT (The recovery test is analyzed using the Theis solution)
}
\examples{
coeffs <- stehfest_coeffs(8)
td <- logseq(-2,4,50)
W <- stehfest_inversion(td, coeffs, agarwal_recovery_WF_LT)
plot(td, W, log = "xy")
}
\references{
Agarwal, R. (1980). A New Method To Account For Producing Time Effects When Drawdown
Type Curves Are Used To Analyze Pressure Buildup And Other Test Data. SPE Paper 9289.
}
\seealso{
Other agarwal_recovery functions: \code{\link{agarwal_recovery_calculate_parameters}},
  \code{\link{agarwal_recovery_solution_dlogt}},
  \code{\link{agarwal_recovery_solution_initial}},
  \code{\link{agarwal_recovery_solution}},
  \code{\link{agarwal_recovery_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{agarwal_recovery functions}
