% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agarwal_skin_utilities.R
\name{agarwal_skin_calculate_parameters}
\alias{agarwal_skin_calculate_parameters}
\title{agarwal_skin_calculate_parameters}
\usage{
agarwal_skin_calculate_parameters(ptest, par, hydraulic = TRUE)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{par}{A list with the values of slope and intercept (a and t0) of the straight
line fitted to the drawdown data using the Cooper-Jacob approach, and the value of
skin factor(sigma)}

\item{hydraulic}{Logical flag to indicate if hydraulic parameters are calculated. If False,
the the statistcal parameter (a and t0) are calculated.}
}
\value{
A list with
\itemize{
\item Tr: Transmissivity
\item Ss: Storage coefficient
\item cd: Wellbore storage
\item rd: Dimensionless radius
}
}
\description{
Function to calculate the hydraulic parameters for Agarwal solution (skin and wellbore storage)
}
\examples{
data(papadopulos_cooper)
}
\references{
Agarwal, R.; Al-Hussainy, R. & Ramey, H. An Investigation of Wellbore Storage and
Skin Effect in Unsteady Liquid Flow: I. Analytical Treatment. Society of Petroleum
Engineers Journal., 1970
}
\seealso{
Other agarwal_skin functions: \code{\link{agarwal_skin_WF_LT}},
  \code{\link{agarwal_skin_solution_dlogt}},
  \code{\link{agarwal_skin_solution_initial}},
  \code{\link{agarwal_skin_solution}},
  \code{\link{agarwal_skin_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@mgmail.com}
}
\concept{agarwal_skin functions}
