% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agarwal_skin_utilities.R
\name{agarwal_skin_solution_initial}
\alias{agarwal_skin_solution_initial}
\title{agarwal_skin_solution_initial}
\usage{
agarwal_skin_solution_initial(ptest)
}
\arguments{
\item{ptest}{A pumping_test object}
}
\value{
A list with :
\itemize{
\item a: Slope of the straight line fitted to the drawdown data using the
Cooper-Jacob approach
\item t0: Intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach
\item sigma: skin factor
}
}
\description{
Function to calculate the initial solution of the Agarwal solution (skin and wellbore storage)
}
\examples{
data(papadopulos_cooper)
}
\references{
Agarwal, R.; Al-Hussainy, R. & Ramey, H. An Investigation of Wellbore Storage and
Skin Effect in Unsteady Liquid Flow: I. Analytical Treatment. Society of Petroleum
Engineers Journal., 1970
}
\seealso{
Other agarwal_skin functions: \code{\link{agarwal_skin_WF_LT}},
  \code{\link{agarwal_skin_calculate_parameters}},
  \code{\link{agarwal_skin_solution_dlogt}},
  \code{\link{agarwal_skin_solution}},
  \code{\link{agarwal_skin_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{agarwal_skin functions}
