% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agarwal_skin_utilities.R
\name{agarwal_skin_well_function}
\alias{agarwal_skin_well_function}
\title{agarwal_skin_well_function}
\usage{
agarwal_skin_well_function(td, par)
}
\arguments{
\item{td}{Numeric vector with time}

\item{par}{A list with the values of wellbore storage(cd), dimensionless radius (rd)
and skin factor (sigma)}
}
\value{
Numeric vector with the values of dimensionless drawdown
}
\description{
Function to calculate the well function using the Agarwal solution (skin and wellbore storage)
}
\examples{
data(papadopulos_cooper)
}
\references{
Agarwal, R.; Al-Hussainy, R. & Ramey, H. An Investigation of Wellbore Storage and
Skin Effect in Unsteady Liquid Flow: I. Analytical Treatment. Society of Petroleum
Engineers Journal., 1970
}
\seealso{
Other agarwal_skin functions: \code{\link{agarwal_skin_WF_LT}},
  \code{\link{agarwal_skin_calculate_parameters}},
  \code{\link{agarwal_skin_solution_dlogt}},
  \code{\link{agarwal_skin_solution_initial}},
  \code{\link{agarwal_skin_solution}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{agarwal_skin functions}
