% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agarwal_skin_utilities.R
\name{agarwal_skin_WF_LT}
\alias{agarwal_skin_WF_LT}
\title{agarwal_skin_WF_LT}
\usage{
agarwal_skin_WF_LT(p, arg1, arg2, arg3)
}
\arguments{
\item{p}{Laplace Transform parameter}

\item{arg1}{value of wellbore coefficient (cd)}

\item{arg2}{value of dimensionless radius (rd)}

\item{arg3}{value of skin factor (sigma)}
}
\value{
A numeric vector with the calculated drawdown in the Laplace domain
}
\description{
Function to calculate the well function of the Agarwal solution (skin and wellbore storage) in the Laplace domain
}
\examples{
data(papadopulos_cooper)
}
\references{
Agarwal, R.; Al-Hussainy, R. & Ramey, H. An Investigation of Wellbore Storage and
Skin Effect in Unsteady Liquid Flow: I. Analytical Treatment. Society of Petroleum
Engineers Journal., 1970
}
\seealso{
Other agarwal_skin functions: \code{\link{agarwal_skin_calculate_parameters}},
  \code{\link{agarwal_skin_solution_dlogt}},
  \code{\link{agarwal_skin_solution_initial}},
  \code{\link{agarwal_skin_solution}},
  \code{\link{agarwal_skin_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{agarwal_skin functions}
