% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquifer_utilities.R
\name{aquifer}
\alias{aquifer}
\title{aquifer}
\usage{
aquifer(name, solution, type, xlim, ylim, nx, ny, hydraulic.pars)
}
\arguments{
\item{name}{Name of the aquifer}

\item{solution}{Analytical solution to be used}

\item{type}{Type of aquifer to be defined. Currently the supported types include:
\itemize{
\item Infinite
\item semi-infinite: with no-flow or constant head boundaries
\item strip:  with no-flow or constant head boundaries or a combination of both
\item wedge
\item U-shaped
\item rectangular
\item circular}}

\item{xlim}{Numeric vector with limits of the aquifer in the x direction}

\item{ylim}{Numeric vector with limits of the aquifer in the y direction}

\item{nx}{Integer with the number of nodes in the x direction}

\item{ny}{Integer with the number of nodes in the y direction}

\item{hydraulic.pars}{A list with the hydraulic parameters of the aquifer}
}
\value{
This function returns an aquifer object
}
\description{
Function to create an aquifer object
}
\seealso{
Other aquifer functions: \code{\link{add.well<-}},
  \code{\link{calculate_drawdown}},
  \code{\link{print.aquifer}},
  \code{\link{summary.aquifer}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{aquifer functions}
