% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_utilities.R
\name{bandwidth_lpridge}
\alias{bandwidth_lpridge}
\title{bandwidth_lpridge}
\usage{
bandwidth_lpridge(t, s, h.range = c(0.01, 1), seed = 12345)
}
\arguments{
\item{t}{Numeric vector}

\item{s}{Numeric vector}

\item{h.range}{values}

\item{seed}{Random seed}
}
\value{
This function returns a list with the following entries:
\itemize{
\item h: bandwidth values
\item cverror: Crossvalidation error
}
}
\description{
Function to calculate the ridge regression bandwidth
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{smoothing functions}
