% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twalk_utilities.R
\name{blow_kernel}
\alias{blow_kernel}
\title{blow_kernel}
\usage{
blow_kernel(x, xp, phi)
}
\arguments{
\item{x}{Numeric value}

\item{xp}{Numeric value}

\item{phi}{Logical vector with the parameters to change.}
}
\value{
Numeric value
}
\description{
Function to calculate blow step
}
\references{
Christen, J. A. & Fox, C. A general purpose sampling algorithm for continuous distributions (the t-walk).
Bayesian Analysis, 2010, 5, 2, 263-281.
}
\seealso{
Other twalk_auxiliary_function functions: \code{\link{hop_kernel}},
  \code{\link{log_ratio_density_blow}},
  \code{\link{log_ratio_density_hop}}, \code{\link{psi1}},
  \code{\link{run_twalk_MCMC}},
  \code{\link{traverse_kernel}}, \code{\link{walk_kernel}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{twalk_auxiliary_function functions}
