% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest_data.R
\docType{data}
\name{boulton}
\alias{boulton}
\title{estimation of a pumping test in an unconfined aquifer with the Boulton's
Solution}
\format{A data frame with 132 rows and 2 variables:
\describe{
  \item{t}{time, in seconds}
  \item{s}{drawdown, in meters}
}}
\source{
G. de Marsily, cours DEA Paris 6, "Aquifere cotier de Nefza, Tunisie".
Piezometer A3
}
\usage{
boulton
}
\description{
The information of this pumping test is:
\itemize{
\item Pumping rate (in m3/s): Q <- 0.030
\item Distance to pumping well (in m): r <- 20
}
}
\details{
estimation with Boulton (1963) model for unconfined aquifer of a pumping
test in coastal aquifer
}
\keyword{datasets}
