% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boulton_utilities.R
\name{boulton_calculate_parameters}
\alias{boulton_calculate_parameters}
\title{boulton_calculate_parameters}
\usage{
boulton_calculate_parameters(ptest, par, hydraulic = TRUE)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{par}{List with the parameters a (slope), t0 and t1 are the intercepts of the
straight line fitted to the drawdown using the Cooper-Jacob method for early and late
times respectively.}

\item{hydraulic}{Logical flag to indicate if hydraulic parameters are calculated. If False,
the the statistcal parameter are calculated.}
}
\value{
A list with the values of transmisivity (Tr), storage coefficient(Ss), Drainage Porosity
(omegad) and radius of influence (radius_influence)
}
\description{
Function to calculate the hydraulic parameters of the Boulton model
}
\examples{
data(boulton)
ptest <- pumping_test("Well1", Q = 0.03, r = 20, t = boulton$t, s = boulton$s)
boulton_sol0 <- boulton_solution_initial(ptest)
boulton.par <- boulton_calculate_parameters(ptest, boulton_sol0)
}
\references{
Boulton, N. The drawdown of the water-table under non-steady conditions near a pumped
well in an unconfined formation. Proceedings of the Institution of Civil Engineers,
1954, 3, 564-579
}
\seealso{
Other boulton functions: \code{\link{boulton_WF_LT}},
  \code{\link{boulton_solution_dlogt}},
  \code{\link{boulton_solution}},
  \code{\link{boulton_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{boulton functions}
