% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boulton_utilities.R
\name{boulton_solution}
\alias{boulton_solution}
\title{boulton_solution}
\usage{
boulton_solution(ptest, a, t0, t1, phi, t)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{a}{Slope of the straight line fitted to the drawdown data using the
Cooper-Jacob approach}

\item{t0}{Intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach (early time)}

\item{t1}{Intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach (late time)}

\item{phi}{Delay parameter. Dimensionless parameter defined as
\deqn{\phi = \frac{\alpha_{1} r^{2} S}{T}}
where \eqn{\alpha_{1}} is a fitting parameter without a physical estimation, \eqn{r} is
the distance between the pumping and observation well, \eqn{S} is the storage coefficient and
\eqn{T} is the transmissivity.}

\item{t}{Numeric vector with the time values}
}
\value{
A vector with the calculated drawdown
}
\description{
Function to calculate the drawdown of a Boulton model
}
\examples{
data(boulton)
ptest <- pumping_test("Well1", Q = 0.03, r = 20, t = boulton$t, s = boulton$s)
boulton_sol0 <- boulton_solution_initial(ptest)
boulton_sol1 <- boulton_solution(ptest, boulton_sol0$a, boulton_sol0$t0,
                boulton_sol0$t1, boulton_sol0$phi, boulton$t)
}
\references{
Boulton, N. The drawdown of the water-table under non-steady conditions near a pumped
well in an unconfined formation. Proceedings of the Institution of Civil Engineers,
1954, 3, 564-579
}
\seealso{
Other boulton functions: \code{\link{boulton_WF_LT}},
  \code{\link{boulton_calculate_parameters}},
  \code{\link{boulton_solution_dlogt}},
  \code{\link{boulton_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{boulton functions}
