% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boulton_utilities.R
\name{boulton_solution_dlogt}
\alias{boulton_solution_dlogt}
\title{boulton_solution_dlogt}
\usage{
boulton_solution_dlogt(ptest, a, t0, t1, phi, t)
}
\arguments{
\item{ptest}{value}

\item{a}{Slope of the straight line fitted to the drawdown data using the
Cooper-Jacob approach}

\item{t0}{Intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach (early time)}

\item{t1}{Intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach (late time)}

\item{phi}{Delay parameter. Dimensionless parameter defined as
\deqn{\phi = \frac{\alpha_{1} r^{2} S}{T}}
where \eqn{\alpha_{1}} is a fitting parameter without a physical estimation, \eqn{r} is
the distance between the pumping and observation well, \eqn{S} is the storage coefficient and
\eqn{T} is the transmissivity.}

\item{t}{value}
}
\value{
This function returns
}
\description{
Function to calculate the derivative of the drawdown with respect to the
log of time using the Boulton solution
}
\examples{
data(boulton)
ptest <- pumping_test('Well1', Q = 0.03, r = 20, t = boulton$t, s = boulton$s)
boulton.sol <- boulton_solution_initial(ptest)
boulton.dsol <- boulton_solution_dlogt(ptest, boulton.sol$a, boulton.sol$t0,
                boulton.sol$t1, boulton.sol$phi, boulton$t)
}
\references{
Boulton, N. The drawdown of the water-table under non-steady conditions near a pumped
well in an unconfined formation. Proceedings of the Institution of Civil Engineers,
1954, 3, 564-579
}
\seealso{
Other boulton functions: \code{\link{boulton_WF_LT}},
  \code{\link{boulton_calculate_parameters}},
  \code{\link{boulton_solution}},
  \code{\link{boulton_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{boulton functions}
