% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boulton_utilities.R
\name{boulton_well_function}
\alias{boulton_well_function}
\title{boulton_well_function}
\usage{
boulton_well_function(td, par)
}
\arguments{
\item{td}{Numeric dimensionless time}

\item{par}{list with the delay parameter (phi), the sigma parameter and the coeffs vector}
}
\value{
Value of the well function of the Boulton model
}
\description{
Well function of the Boulton model for the phreatic aquifer
}
\examples{
td <- logseq(-2, 4, 30)
par <- list(sigma =0.1, phi = 0.1)
W.boulton <- boulton_well_function(td, par)
plot(1/td, W.boulton, type = "p", log = "xy")
}
\references{
Boulton, N. The drawdown of the water-table under non-steady conditions near a pumped
well in an unconfined formation. Proceedings of the Institution of Civil Engineers,
1954, 3, 564-579
}
\seealso{
Other boulton functions: \code{\link{boulton_WF_LT}},
  \code{\link{boulton_calculate_parameters}},
  \code{\link{boulton_solution_dlogt}},
  \code{\link{boulton_solution}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{boulton functions}
