% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boulton_utilities.R
\name{boulton_WF_LT}
\alias{boulton_WF_LT}
\title{boulton_WF_LT}
\usage{
boulton_WF_LT(p, arg1, arg2)
}
\arguments{
\item{p}{Laplace parameter}

\item{arg1}{Value of sigma parameter defined as
\deqn{\sigma = \frac{S}{S_{y}} }}

\item{arg2}{Value of the delay parameter(phi):
\deqn{\phi = \frac{\alpha_{1} r^{2} S}{T}}}
}
\value{
The value of the drawdown in the Laplace domain
}
\description{
Function to calculate the Boulton solution in the Laplace domain.
}
\examples{
sigma <- 0.1
phi <- 0.1
p <- 0.5
res <- boulton_WF_LT(p, sigma, phi)
}
\references{
Boulton, N. The drawdown of the water-table under non-steady conditions near a pumped
well in an unconfined formation. Proceedings of the Institution of Civil Engineers,
1954, 3, 564-579
}
\seealso{
Other boulton functions: \code{\link{boulton_calculate_parameters}},
  \code{\link{boulton_solution_dlogt}},
  \code{\link{boulton_solution}},
  \code{\link{boulton_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{boulton functions}
