% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bouwer_rice_utilities.R
\name{bouwer_rice_initial_solution}
\alias{bouwer_rice_initial_solution}
\title{bouwer_rice_initial_solution}
\usage{
bouwer_rice_initial_solution(ptest)
}
\arguments{
\item{ptest}{A pumping_test object}
}
\value{
A list with the value of t0
}
\description{
Function to calculate the initial value of t0 that defines the Bower-Rice solution
}
\seealso{
Other bouwer_rice functions: \code{\link{Acoeff}},
  \code{\link{Bcoeff}}, \code{\link{Ccoeff}},
  \code{\link{bouwer_rice_solution}},
  \code{\link{bouwer_rice_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{bouwer_rice functions}
