% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bouwer_rice_utilities.R
\name{bouwer_rice_solution}
\alias{bouwer_rice_solution}
\title{bower_rice_solution}
\usage{
bouwer_rice_solution(ptest, t0, t)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{t0}{A numeric value with the t0}

\item{t}{A numeric vector with the time}
}
\value{
A numeric vector with the calculated drawdown
}
\description{
Function to calculate the drawdown during a slug test using the Bower-Rice solution.
}
\seealso{
Other bouwer_rice functions: \code{\link{Acoeff}},
  \code{\link{Bcoeff}}, \code{\link{Ccoeff}},
  \code{\link{bouwer_rice_initial_solution}},
  \code{\link{bouwer_rice_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{bouwer_rice functions}
