% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bouwer_rice_utilities.R
\name{bouwer_rice_well_function}
\alias{bouwer_rice_well_function}
\title{bouwer_rice_well_function}
\usage{
bouwer_rice_well_function(td, par)
}
\arguments{
\item{td}{A numeric vector with the dimensionless time}

\item{par}{A list with the value of the t0}
}
\value{
A numeric vector with the normalized drawdown
}
\description{
Function to calculate the well function of the Bower-Rice solution
}
\seealso{
Other bouwer_rice functions: \code{\link{Acoeff}},
  \code{\link{Bcoeff}}, \code{\link{Ccoeff}},
  \code{\link{bouwer_rice_initial_solution}},
  \code{\link{bouwer_rice_solution}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{bouwer_rice functions}
