% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_function_utilities.R
\name{box_cox_transf}
\alias{box_cox_transf}
\title{box_cox_transform}
\usage{
box_cox_transf(y, lambda)
}
\arguments{
\item{y}{A numeric vector with the values to be transformed.}

\item{lambda}{A numeric value with the lambda value that defines the transformation.}
}
\value{
A numeric vector with the transformed values
}
\description{
Function to apply the Box-Cox transform to a vector.
}
\seealso{
Other objective_function functions: \code{\link{loglikelihood_bc}},
  \code{\link{loglikelihood}},
  \code{\link{max_absolute_deviation}},
  \code{\link{mean_absolute_deviation}},
  \code{\link{mean_square_error}},
  \code{\link{residual_sum_squares}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{objective_function functions}
