% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquifer_utilities.R
\name{calculate_drawdown}
\alias{calculate_drawdown}
\title{calculate_drawdown}
\usage{
calculate_drawdown(aq, current.times)
}
\arguments{
\item{aq}{An aquifer object}

\item{current.times}{A numeric vector with the current times}
}
\value{
This function returns a matrix where the calculated drawdowns are assigned in 
the columns
}
\description{
Function to calculate the drawdown produced by well withdrawals at given times
}
\seealso{
Other aquifer functions: \code{\link{add.well<-}},
  \code{\link{aquifer}}, \code{\link{print.aquifer}},
  \code{\link{summary.aquifer}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{aquifer functions}
