% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bouwer_rice_utilities.R
\name{Ccoeff}
\alias{Ccoeff}
\title{Ccoeff}
\usage{
Ccoeff(x)
}
\arguments{
\item{x}{A numeric value}
}
\value{
This function returns the value of the C coefficient
}
\description{
Function to calculate the C coefficient of the Bouwer and Rice solution.
}
\seealso{
Other bouwer_rice functions: \code{\link{Acoeff}},
  \code{\link{Bcoeff}},
  \code{\link{bouwer_rice_initial_solution}},
  \code{\link{bouwer_rice_solution}},
  \code{\link{bouwer_rice_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{bouwer_rice functions}
