% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest.R
\name{confint.pumping_test}
\alias{confint.pumping_test}
\title{confint.pumping_test}
\usage{
\method{confint}{pumping_test}(object, parm, level = 0.95, ...,
  slug = FALSE, method = "wald", d = 20, neval = 100,
  seed = 12345)
}
\arguments{
\item{object}{An estimated pumping_test object}

\item{parm}{Name of parameters to be analyzed}

\item{level}{A numeric vector with the probabilities used to calculate the confidence
intervals.}

\item{...}{additional argument(s) for methods.}

\item{slug}{Logical flag to indicate a slug test.}

\item{method}{A character string with the method to estimate the confidence intervals.
The methods currently supported are:
\itemize{
\item wald
\item bootstrap
\item jackniffe
}.}

\item{d}{Number of points used in the calculation of the derivative.}

\item{neval}{Number of bootstrap realizations used in the estimation of confidence
intervals.}

\item{seed}{Random seed.}
}
\value{
In the case of method=wald the function returns a list with the following entries:
\itemize{
\item hydraulic.parameters.names: A character vector with the names of the hydraulic parameters
\item hydraulic.parameters.ci: Confidence intervals of the hydraulic parameters
}
In the case of method=bootstrap the function returns a list with the following entries:
\itemize{
\item fit.parameters: A matrix with the values of the parameters of the nonlinear
regression model
\item total.drawdown: A matrix with the values of the drawdown
\item total_dsdlogt: A matrix with the values of the derivative of drawdown
\item sci: A matrix with the confidence intervals of drawdown
\item dsci: A matrix with the confidence intervals of the derivative of drawdown
\item hydraulic.parameters: A matrix with the hydraulic parameter estimated from the boostrapp
realizations
\item parameter.names: A vector with the name of the fitting parameters
\item hydraulic.parameters.names: A character vector with the names of the hydraulic parameters
\item hydraulic.parameters.ci: Confidence intervals of the hydraulic parameters
}
}
\description{
Calculates the confindence intervals of the hydraulic parameters, drawdown and
its derivative by bootstrapping the residuals
}
\examples{
data(theis)
ptest.theis <- pumping_test("Well1", Q = 1.388e-2, r = 250, t= theis$t, s = theis$s)
ptest.theis.fit <- fit(ptest.theis, "theis")
hydraulic.parameters(ptest.theis) <- ptest.theis.fit$hydraulic_parameters
fit.parameters(ptest.theis) <- ptest.theis.fit$parameters
model(ptest.theis) <- "theis"
estimated(ptest.theis) <- TRUE
ptest.theis.ci1 <- confint(ptest.theis, level = 0.975)
ptest.theis.ci2 <- confint(ptest.theis, level = c(0.025, 0.975), method = 'bootstrap',
 d = 20, neval = 100)
}
\seealso{
Other base functions: \code{\link{additional.parameters<-}},
  \code{\link{confint_bootstrap}},
  \code{\link{confint_jackniffe}},
  \code{\link{confint_wald}}, \code{\link{estimated<-}},
  \code{\link{evaluate}}, \code{\link{fit.optimization}},
  \code{\link{fit.parameters<-}},
  \code{\link{fit.sampling}}, \code{\link{fit}},
  \code{\link{hydraulic.parameter.names<-}},
  \code{\link{hydraulic.parameters<-}},
  \code{\link{model.parameters}}, \code{\link{model<-}},
  \code{\link{plot.pumping_test}},
  \code{\link{plot_model_diagnostic}},
  \code{\link{plot_sample_influence}},
  \code{\link{plot_uncert}},
  \code{\link{print.pumping_test}},
  \code{\link{pumping_test}}, \code{\link{simulate}},
  \code{\link{summary.pumping_test}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{base functions}
