% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest.R
\name{confint_bootstrap}
\alias{confint_bootstrap}
\title{confint_bootstrap}
\usage{
confint_bootstrap(ptest, level = c(0.025, 0.975), slug = F, d = 20,
  neval = 100, seed = 12345)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{level}{A numeric vector with the significance levels of the confidence intervals.}

\item{slug}{A logical flag indicating if a slug test is used in the analysis.}

\item{d}{An integer value specifying the number of points used in the calculation of the
derivative.}

\item{neval}{An integer value specifying the number of realizations of the resampling procedure.}

\item{seed}{random seed}
}
\value{
This function returns a list with the following entries:
\itemize{
\item fit.parameters: A matrix with the values of the parameters of the nonlinear
regression model
\item method: A character string specifying the method used in the CI estimation (bootstrap)
\item total.drawdown: A matrix with the values of the drawdown
\item total_dsdlogt: A matrix with the values of the derivative of drawdown
\item sci: A matrix with the confidence intervals of drawdown
\item dsci: A matrix with the confidence intervals of the derivative of drawdown
\item hydraulic.parameters: A matrix with the hydraulic parameter estimated from the boostrapp
realizations
\item parameter.names: A vector with the name of the fitting parameters
\item hydraulic.parameters.names: A character vector with the names of the hydraulic parameters
\item hydraulic.parameters.ci: Confidence intervals of the hydraulic parameters
}
}
\description{
Function to calculate the confidence intervals of the hydraulic parameters using
bootstrapping.
}
\seealso{
Other base functions: \code{\link{additional.parameters<-}},
  \code{\link{confint.pumping_test}},
  \code{\link{confint_jackniffe}},
  \code{\link{confint_wald}}, \code{\link{estimated<-}},
  \code{\link{evaluate}}, \code{\link{fit.optimization}},
  \code{\link{fit.parameters<-}},
  \code{\link{fit.sampling}}, \code{\link{fit}},
  \code{\link{hydraulic.parameter.names<-}},
  \code{\link{hydraulic.parameters<-}},
  \code{\link{model.parameters}}, \code{\link{model<-}},
  \code{\link{plot.pumping_test}},
  \code{\link{plot_model_diagnostic}},
  \code{\link{print.pumping_test}},
  \code{\link{pumping_test}}, \code{\link{simulate}},
  \code{\link{summary.pumping_test}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{base functions}
