% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest_data.R
\docType{data}
\name{cooper}
\alias{cooper}
\title{Cooper Slug Test - Initial solution}
\format{A data frame with 100 rows and 2 variables:
\describe{
  \item{dsmax}{dsmax}
  \item{cdmax}{cdmax}
}}
\usage{
cooper
}
\description{
Cooper Slug Test - Initial solution
}
\keyword{datasets}
