% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooper_utilities.R
\name{cooper_calculate_parameters}
\alias{cooper_calculate_parameters}
\title{cooper_calculate_parameters}
\usage{
cooper_calculate_parameters(ptest, par, hydraulic = TRUE)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{par}{A list with the dimensionless wellbore storage coefficient (cd) and the
Intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach (t0)}

\item{hydraulic}{Logical flag to indicate if hydraulic parameters are calculated. If
False, the the statistcal parameter are calculated.}
}
\value{
A list with
\itemize{
  \item Tr: Transmisivity
  \item Ss: Storage coefficient
  \item radius_influence: radius of influence
}
}
\description{
Function to calculate the hydraulic parameters of a cooper solution for a slug test
}
\examples{
data(cooper_slug)
ptest <- pumping_test("Well1", Q = 0, r = 0, t = cooper_slug$t, s = cooper_slug$s)
par <- list(rw = 0.071, rc = 0.025)
ptest$additional_parameters <- par
sol0 <- cooper_solution_initial(ptest)
hydr.par <- cooper_calculate_parameters(ptest,sol0)
print(hydr.par)
}
\references{
Cooper, H. H.; Bredehoeft, J. D. & Papadopulos, I. S. Response of a finite-diameter
well to an instantaneous charge of water Water Resources Research, 1967, 3, 263-269
}
\seealso{
Other cooper functions: \code{\link{cooper_WF_LT}},
  \code{\link{cooper_solution_dlogt}},
  \code{\link{cooper_solution_initial}},
  \code{\link{cooper_solution}},
  \code{\link{cooper_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{cooper functions}
