% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooper_jacob_utilities.R
\name{cooper_jacob_calculate_parameters}
\alias{cooper_jacob_calculate_parameters}
\title{cooper_jacob_calculate_parameters}
\usage{
cooper_jacob_calculate_parameters(ptest, par, hydraulic = TRUE)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{par}{list with the value of a (slope) and t0 (Intercept) of the straight line
fitted to the drawdown data}

\item{hydraulic}{Logical flag to indicate if hydraulic parameters are calculated. If
False, the the statistcal parameter are calculated.}
}
\value{
A list with the transmissivity, storage coefficient and radious of influence
}
\description{
Function to calculate the transmissivity, storage coefficient and radius of influence
using the Cooper-Jacob solution
}
\examples{
data(theis)
ptest <- pumping_test("Well1", Q= 1.3e-3, r = 200, t = theis$t, s = theis$s)
cooper_jacob_sol0 <- cooper_jacob_solution_initial(ptest)
par <- cooper_jacob_calculate_parameters(ptest, cooper_jacob_sol0)
print(par)
}
\references{
Cooper, H. & Jacob, C. A generalized graphical method for evaluating formation
constants and summarizing well field history. Transactions of the American
Geophysical Union, 1946, 27, 526-534
}
\seealso{
Other cooper_jacob functions: \code{\link{cooper_jacob_solution_dlogt}},
  \code{\link{cooper_jacob_solution_initial}},
  \code{\link{cooper_jacob_solution}},
  \code{\link{cooper_jacob_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{cooper_jacob functions}
