% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooper_jacob_utilities.R
\name{cooper_jacob_solution}
\alias{cooper_jacob_solution}
\title{cooper_jacob_solution}
\usage{
cooper_jacob_solution(ptest, a, t0, t)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{a}{Slope of the straight line fitted to the drawdown data using the
Cooper-Jacob approach}

\item{t0}{Intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach}

\item{t}{Numeric vector with the values of time}
}
\value{
This function returns a numeric vector with the calculated drawdown
}
\description{
Function to calculate the drawdown using the Cooper-Jacob solution for a confined aquifer
}
\examples{
data(theis)
ptest <- pumping_test("Well1", Q= 1.3e-3, r = 200, t = theis$t, s = theis$s)
sol0 <- cooper_jacob_solution_initial(ptest)
sol <- cooper_jacob_solution(ptest, sol0$a, sol0$t0, ptest$t)
print(sol)
}
\references{
Cooper, H. & Jacob, C. A generalized graphical method for evaluating formation
constants and summarizing well field history. Transactions of the American
Geophysical Union, 1946, 27, 526-534
}
\seealso{
Other cooper_jacob functions: \code{\link{cooper_jacob_calculate_parameters}},
  \code{\link{cooper_jacob_solution_dlogt}},
  \code{\link{cooper_jacob_solution_initial}},
  \code{\link{cooper_jacob_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{cooper_jacob functions}
