% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooper_jacob_utilities.R
\name{cooper_jacob_solution_initial}
\alias{cooper_jacob_solution_initial}
\title{cooper_jacob_solution_initial}
\usage{
cooper_jacob_solution_initial(ptest)
}
\arguments{
\item{ptest}{A pumping_test object}
}
\value{
A list with the values of a (slope) and t0 (intercept) of the straight line fitted to the
drawdown data
}
\description{
Function to calculate the initial values of the Cooper-Jacob Solution
}
\examples{
data(theis)
ptest <- pumping_test("Well1", Q= 1.3e-3, r = 200, t = theis$t, s = theis$s)
cooper_jacob_sol0 <- cooper_jacob_solution_initial(ptest)
print(cooper_jacob_sol0)
}
\references{
Cooper, H. & Jacob, C. A generalized graphical method for evaluating formation
constants and summarizing well field history. Transactions of the American
Geophysical Union, 1946, 27, 526-534
}
\seealso{
Other cooper_jacob functions: \code{\link{cooper_jacob_calculate_parameters}},
  \code{\link{cooper_jacob_solution_dlogt}},
  \code{\link{cooper_jacob_solution}},
  \code{\link{cooper_jacob_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{cooper_jacob functions}
