% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooper_jacob_utilities.R
\name{cooper_jacob_well_function}
\alias{cooper_jacob_well_function}
\title{cooper_jacob_well_function}
\usage{
cooper_jacob_well_function(u)
}
\arguments{
\item{u}{dimensionless variable u}
}
\value{
This function returns the dimensionless drawdown using the Cooper-Jacob solution
}
\description{
Function to calculate the well function of the Cooper-Jacob solution
}
\examples{
td <- logseq(-2, 4, 50)
u <- 1.0/td
res <- cooper_jacob_well_function(u)
plot(u,res, type="p", log="xy")
}
\references{
Cooper, H. & Jacob, C. A generalized graphical method for evaluating formation
constants and summarizing well field history. Transactions of the American
Geophysical Union, 1946, 27, 526-534
}
\seealso{
Other cooper_jacob functions: \code{\link{cooper_jacob_calculate_parameters}},
  \code{\link{cooper_jacob_solution_dlogt}},
  \code{\link{cooper_jacob_solution_initial}},
  \code{\link{cooper_jacob_solution}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{cooper_jacob functions}
