% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest_data.R
\docType{data}
\name{cooper_slug}
\alias{cooper_slug}
\title{estimation of a slug test with the Cooper Solution}
\format{A data frame with 69 rows and 2 variables:
\describe{
  \item{t}{time, in seconds}
  \item{s}{drawdown, in meters}
}}
\source{
Butler James, 1998, The design, performance and analysis of slug tests.
Lewis Publisher. Table 5.2 pp. 61-62.
}
\usage{
cooper_slug
}
\description{
The information for this slug test is:
\itemize{
\item Radius of the well (in m): rw <- 0.071
\item Radius of the casing (in m): rc <- 0.025
}
}
\details{
Localisation: Test performed in a monitoring well in Lincoln County, Kansas
Geology. The well was screened in a deltaic sequence consisting of mudstone
interbedded with very fine sandstone.
}
\keyword{datasets}
