% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooper_utilities.R
\name{cooper_solution}
\alias{cooper_solution}
\title{cooper_solution}
\usage{
cooper_solution(ptest, cd, t0, t)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{cd}{Value of the dimensionless wellbore storage coefficient}

\item{t0}{Intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach}

\item{t}{Numeric vector with the value of time}
}
\value{
A numeric vector with the calculated drawdown
}
\description{
Function to calculate the drawdown using the Cooper solution for a slug test
}
\examples{
data(cooper_slug)
ptest <- pumping_test("Well1", Q = 0, r = 0, t = cooper_slug$t, s = cooper_slug$s)
sol0 <- cooper_solution_initial(ptest)
sol <- cooper_solution(ptest, sol0$cd, sol0$t0, ptest$t)
print(sol)
}
\references{
Cooper, H. H.; Bredehoeft, J. D. & Papadopulos, I. S. Response of a finite-diameter
well to an instantaneous charge of water Water Resources Research, 1967, 3, 263-269
}
\seealso{
Other cooper functions: \code{\link{cooper_WF_LT}},
  \code{\link{cooper_calculate_parameters}},
  \code{\link{cooper_solution_dlogt}},
  \code{\link{cooper_solution_initial}},
  \code{\link{cooper_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{cooper functions}
