% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooper_utilities.R
\name{cooper_solution_dlogt}
\alias{cooper_solution_dlogt}
\title{cooper_solution_dlogt}
\usage{
cooper_solution_dlogt(ptest, cd, t0, t)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{cd}{Value of the dimensionless storage coefficient (cd)}

\item{t0}{Intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach}

\item{t}{A numeric vector with the time values}
}
\value{
A numeric vector with the values of the drawdown
}
\description{
Function to calculate the derivative of the drawdown with respect to the log of time for the
Cooper solution for slug tests
}
\examples{
data(cooper_slug)
ptest <- pumping_test("Well1", Q = 0, r = 0, t = cooper_slug$t, s = cooper_slug$s)
sol0 <- cooper_solution_initial(ptest)
sol_dlogt <- cooper_solution_dlogt(ptest, sol0$cd, sol0$t0, ptest$t)
print(sol_dlogt)
}
\references{
Cooper, H. H.; Bredehoeft, J. D. & Papadopulos, I. S. Response of a finite-diameter
well to an instantaneous charge of water Water Resources Research, 1967, 3, 263-269
}
\seealso{
Other cooper functions: \code{\link{cooper_WF_LT}},
  \code{\link{cooper_calculate_parameters}},
  \code{\link{cooper_solution_initial}},
  \code{\link{cooper_solution}},
  \code{\link{cooper_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{cooper functions}
