% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooper_utilities.R
\name{cooper_solution_initial}
\alias{cooper_solution_initial}
\title{cooper_solution_initial}
\usage{
cooper_solution_initial(ptest)
}
\arguments{
\item{ptest}{A pumping_test object}
}
\value{
A list with
\itemize{
  \item cd: Initial value of the wellbore storage coefficient
  \item t0: Intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach
}
}
\description{
Function to calculate the initial values of the Cooper solution
}
\examples{
data(cooper_slug)
ptest <- pumping_test("Well1", Q = 0.0, r = 0.0, t = cooper_slug$t, s = cooper_slug$s)
sol0 <- cooper_solution_initial(ptest)
print(sol0)
}
\references{
Cooper, H. H.; Bredehoeft, J. D. & Papadopulos, I. S. Response of a finite-diameter
well to an instantaneous charge of water Water Resources Research, 1967, 3, 263-269
}
\seealso{
Other cooper functions: \code{\link{cooper_WF_LT}},
  \code{\link{cooper_calculate_parameters}},
  \code{\link{cooper_solution_dlogt}},
  \code{\link{cooper_solution}},
  \code{\link{cooper_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{cooper functions}
