% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooper_utilities.R
\name{cooper_well_function}
\alias{cooper_well_function}
\title{cooper_well_function}
\usage{
cooper_well_function(td, par)
}
\arguments{
\item{td}{Numeric vector with dimensionless time}

\item{par}{List with the stehfest coefficients and cd (dimensionless storage coefficient)}
}
\value{
A numeric vector with the value of the Cooper well function, that is, the dimensionless drawdown
}
\description{
Function to calculate the well function of the cooper solution for slug test
}
\examples{
td <- logseq(-2, 4, 50)
par <- list(cd = 1, rw = 1)
res <- cooper_well_function(td, par)
}
\references{
Cooper, H. H.; Bredehoeft, J. D. & Papadopulos, I. S. Response of a finite-diameter
well to an instantaneous charge of water Water Resources Research, 1967, 3, 263-269
}
\seealso{
Other cooper functions: \code{\link{cooper_WF_LT}},
  \code{\link{cooper_calculate_parameters}},
  \code{\link{cooper_solution_dlogt}},
  \code{\link{cooper_solution_initial}},
  \code{\link{cooper_solution}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{cooper functions}
