% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooper_utilities.R
\name{cooper_WF_LT}
\alias{cooper_WF_LT}
\title{cooper_WF_LT}
\usage{
cooper_WF_LT(p, arg1, arg2)
}
\arguments{
\item{p}{Laplace parameter}

\item{arg1}{Value of the dimensionless wellbore storage coefficient (cd)}

\item{arg2}{Value or the dimensionless radius (rd)}
}
\value{
The value of the Cooper's solution for slug test in the Laplace domain
}
\description{
Cooper's well function for slug tests in the Laplace domain
}
\examples{
p <- seq(1,10,0.1)
res <- cooper_WF_LT(p, 0.1, 0.1)
print(res)
}
\references{
Cooper, H. H.; Bredehoeft, J. D. & Papadopulos, I. S. Response of a finite-diameter
well to an instantaneous charge of water Water Resources Research, 1967, 3, 263-269
}
\seealso{
Other cooper functions: \code{\link{cooper_calculate_parameters}},
  \code{\link{cooper_solution_dlogt}},
  \code{\link{cooper_solution_initial}},
  \code{\link{cooper_solution}},
  \code{\link{cooper_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{cooper functions}
