% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest.R
\name{evaluate}
\alias{evaluate}
\title{evaluate}
\usage{
evaluate(ptest, model, test.time = TRUE, n = 50)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{model}{Character string specifying the model used to calculate the drawdown}

\item{test.time}{A logical flag indicating if the evaluation of the pumping test data needs to
be calculated at the drawdown times}

\item{n}{Number of points used to calculate the drawdown if test.time = FALSE}
}
\value{
A list with the entries
\itemize{
\item t: Vector with the time
\item s: Vector with the calculated drawdown
\item dsdlnt: Vector with the derivative of drawdown with respect to the log of time
}
}
\description{
Function to calculate the drawdown in time or space from a set of parameters determined using
the \code{fit} function.
}
\examples{
data(theis)
ptest <- pumping_test("Well1", Q = 1.3e-3, r = 200, t = theis$t, s = theis$s)
ptest.fit <- fit(ptest, "theis")
hydraulic.parameters(ptest) <- ptest.fit$hydraulic_parameters
fit.parameters(ptest) <- ptest.fit$parameters
model(ptest) <- "theis"
estimated(ptest) <- TRUE
res <- evaluate(ptest, "theis")
}
\seealso{
Other base functions: \code{\link{additional.parameters<-}},
  \code{\link{confint.pumping_test}},
  \code{\link{confint_bootstrap}},
  \code{\link{confint_jackniffe}},
  \code{\link{confint_wald}}, \code{\link{estimated<-}},
  \code{\link{fit.optimization}},
  \code{\link{fit.parameters<-}},
  \code{\link{fit.sampling}}, \code{\link{fit}},
  \code{\link{hydraulic.parameter.names<-}},
  \code{\link{hydraulic.parameters<-}},
  \code{\link{model.parameters}}, \code{\link{model<-}},
  \code{\link{plot.pumping_test}},
  \code{\link{plot_model_diagnostic}},
  \code{\link{plot_sample_influence}},
  \code{\link{plot_uncert}},
  \code{\link{print.pumping_test}},
  \code{\link{pumping_test}}, \code{\link{simulate}},
  \code{\link{summary.pumping_test}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{base functions}
