% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theis_utilities.R
\name{exponential_integral}
\alias{exponential_integral}
\title{exponential_integral}
\usage{
exponential_integral(x)
}
\arguments{
\item{x}{Numeric value}
}
\value{
This function returns the value of the exponential integral
}
\description{
Function to calculate the exponential integral used in the Theis well function
}
\examples{
x <- seq(0.01,10,0.05)
Ei <- exponential_integral(x)
plot(x, Ei, type = "p", log = 'y')
}
\seealso{
Other theis functions: \code{\link{theis_WF_LT_dlogt}},
  \code{\link{theis_WF_LT}},
  \code{\link{theis_calculate_parameters}},
  \code{\link{theis_solution_dlogt}},
  \code{\link{theis_solution_initial}},
  \code{\link{theis_solution}},
  \code{\link{theis_well_function_large}},
  \code{\link{theis_well_function_small}},
  \code{\link{theis_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{theis functions}
