% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest.R
\name{fit.optimization}
\alias{fit.optimization}
\title{fit.optimization}
\usage{
fit.optimization(ptest, model, obj.fn = "rss", opt.method = "nls",
  lower = 1e-09, upper = Inf, control.par, seed = 12345)
}
\arguments{
\item{ptest}{A pumping_test object.}

\item{model}{A character string specifying the model used in the parameter estimation.}

\item{obj.fn}{A character string specifying the objective function used in the parameter estimation.
Currently the following objective functions are included:
\itemize{
\item 'rss': Residual sum of squares
\item 'mnad': Mean absolute deviation
\item 'mxad': Maximum absolute deviation
\item 'loglik': Loglikelihood function
}}

\item{opt.method}{A character string specifying the optimization method used in the parameter estimation.
Currently the following methologies are included:
\itemize{
\item 'nls': Nonlinear regression
\item 'sa': Simulated Annealing (GenSA package)
\item 'ga': Genetic Algorithms (GA package)
\item 'l-bfgs-b': using optim function (stats package)
\item 'pso': Particle Swarm Optimization (pso package)
\item 'copulaedas': Estimation of Distribution Algorithms Based on Copulas (copulaedas package)
\item 'de': Differential Evolution (DEoptim package)
}}

\item{lower}{A numeric vector with the lower values of the search region}

\item{upper}{A numeric vector with the upper values of the search region}

\item{control.par}{A list with the parameters of the optimization method}

\item{seed}{A random seed}
}
\value{
A list with the following entries:
\itemize{
\item hydraulic_parameters: hydraulic parameters of the model (includes transmissivity, storage
coefficient and radius of influence, or others)
\item parameters: fitted parameters (including a and t0 and other depending on the model)
\item resfit: The list or object returned by the optimization driver of each method.
\item value: The value of the objective function reached at the end of the optimization run.
}
}
\description{
Function to estimate the aquifer parameters from a pumping test using several optimization
functions.
}
\examples{
\dontrun{
# Define pumping_test object
data("boulton")
ptest.boulton <- pumping_test("Well1", Q = 0.03, r = 20,
                              t = boulton$t, s = boulton$s)
# Parameter estimation using L-BFGS-B
ptest.boulton.bfgs.rss <- fit.optimization(ptest.boulton,
"boulton", obj.fn = "rss", opt.method = "l-bfgs-b",
seed = 54321)
# Parameter estimation using Simulated Annealing
ptest.boulton.sa.rss <- fit.optimization(ptest.boulton,
"boulton", obj.fn = "rss", opt.method = "sa", seed = 54321)
# Parameter estimation using Genetic Algorithms
ptest.boulton.ga.rss <- fit.optimization(ptest.boulton,
"boulton", obj.fn = "rss", opt.method = "ga", seed = 54321)
# Parameter estimation using Differential Evolution
ptest.boulton.de.rss <- fit.optimization(ptest.boulton,
                        "boulton", obj.fn = "rss", opt.method = "de", seed = 54321)
# Parameter estimation using Particle Swarm Optimization
ptest.boulton.pso.rss <- fit.optimization(ptest.boulton,
                         "boulton", obj.fn = "rss", opt.method = "pso", seed = 54321)
}
}
\seealso{
Other base functions: \code{\link{additional.parameters<-}},
  \code{\link{confint.pumping_test}},
  \code{\link{confint_bootstrap}},
  \code{\link{confint_jackniffe}},
  \code{\link{confint_wald}}, \code{\link{estimated<-}},
  \code{\link{evaluate}}, \code{\link{fit.parameters<-}},
  \code{\link{fit.sampling}}, \code{\link{fit}},
  \code{\link{hydraulic.parameter.names<-}},
  \code{\link{hydraulic.parameters<-}},
  \code{\link{model.parameters}}, \code{\link{model<-}},
  \code{\link{plot.pumping_test}},
  \code{\link{plot_model_diagnostic}},
  \code{\link{plot_sample_influence}},
  \code{\link{plot_uncert}},
  \code{\link{print.pumping_test}},
  \code{\link{pumping_test}}, \code{\link{simulate}},
  \code{\link{summary.pumping_test}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{base functions}
