% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest.R
\name{fit.parameters<-}
\alias{fit.parameters<-}
\title{fit.parameters<-}
\usage{
fit.parameters(x) <- value
}
\arguments{
\item{x}{A pumping_test object}

\item{value}{A list with the parameters obtained by the fit procedure}
}
\value{
The pumping_test object with the fit parameters
}
\description{
Function to assign the fit parameters (a, t0 and others depending on the model) to a
pumping_test object
}
\examples{
data(theis)
ptest.theis <- pumping_test('Well1', Q = 1.38e-2, r = 250, t = theis$t, s = theis$s)
ptest.theis.fit <- fit(ptest.theis, 'theis')
hydraulic.parameters(ptest.theis) <- ptest.theis.fit$hydraulic_parameters
fit.parameters(ptest.theis) <- ptest.theis.fit$parameters
}
\seealso{
Other base functions: \code{\link{additional.parameters<-}},
  \code{\link{confint.pumping_test}},
  \code{\link{confint_bootstrap}},
  \code{\link{confint_jackniffe}},
  \code{\link{confint_wald}}, \code{\link{estimated<-}},
  \code{\link{evaluate}}, \code{\link{fit.optimization}},
  \code{\link{fit.sampling}}, \code{\link{fit}},
  \code{\link{hydraulic.parameter.names<-}},
  \code{\link{hydraulic.parameters<-}},
  \code{\link{model.parameters}}, \code{\link{model<-}},
  \code{\link{plot.pumping_test}},
  \code{\link{plot_model_diagnostic}},
  \code{\link{print.pumping_test}},
  \code{\link{pumping_test}}, \code{\link{simulate}},
  \code{\link{summary.pumping_test}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{base functions}
