% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest.R
\name{fit}
\alias{fit}
\title{fit}
\usage{
fit(ptest, model, control.par, trace = F)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{model}{Character string specifying the model used in the estimation}

\item{control.par}{A list with parameters of the parameter estimation using nonlinear
regression}

\item{trace}{A logical flag indicating if the results of the nonlinear regression are
printed on the screen}
}
\value{
A list with the following entries:
\itemize{
\item hydraulic_parameters: hydraulic parameters of the model (includes transmissivity, storage
coefficient and radius of influence)
\item paraemters: fitted parameters (includes a and t0)
\item resfit: List with the results of the nonlinear regression
\item value: Value of the residual sum of squares
}
}
\description{
Generic function to estimate the aquifer parameters from a pumping test. This function uses
nonlinear least squares to estimate these parameters.
}
\examples{
#Fit test from confined aquifer
data(theis)
ptest <- pumping_test('Well1', Q = 1.388e-2, r = 250, t = theis$t, s = theis$s)
res_th <- fit(ptest, 'theis')
print(res_th)
#Fit test from confined aquifer
res_cj <- fit(ptest, 'cooper_jacob')
}
\seealso{
Other base functions: \code{\link{additional.parameters<-}},
  \code{\link{confint.pumping_test}},
  \code{\link{confint_bootstrap}},
  \code{\link{confint_jackniffe}},
  \code{\link{confint_wald}}, \code{\link{estimated<-}},
  \code{\link{evaluate}}, \code{\link{fit.optimization}},
  \code{\link{fit.parameters<-}},
  \code{\link{fit.sampling}},
  \code{\link{hydraulic.parameter.names<-}},
  \code{\link{hydraulic.parameters<-}},
  \code{\link{model.parameters}}, \code{\link{model<-}},
  \code{\link{plot.pumping_test}},
  \code{\link{plot_model_diagnostic}},
  \code{\link{print.pumping_test}},
  \code{\link{pumping_test}}, \code{\link{simulate}},
  \code{\link{summary.pumping_test}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{base functions}
