% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest_data.R
\docType{data}
\name{general_radial_flow1}
\alias{general_radial_flow1}
\title{General Radial Flow: Example 1}
\format{A data frame with 78 rows and 2 variables
\describe{
\item{t}{Time in seconds}
\item{s}{Drawdown in m}
}}
\source{
G. Lods and P. Gouze (2004) WTFM, software for well test analysis in
fractured media combining fractional flow with double porosity approaches.
Computers and Geosciences. Vol 30. pp. 937-947 Example CS2 page 943-944
}
\usage{
general_radial_flow1
}
\description{
Data from well well Sha1B. The information of this pumping tests is:
\itemize{
\item Distance to pumping well (in m): r <- 26.2
\item Radius of the well (in m): rw <- 0.1
\item Radius of the casing (in m): rc <- 0.1
\item Pumping rate in(m3/2): Q <- 0.02322
}
}
\keyword{datasets}
