% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_radial_flow_utilities.R
\name{general_radial_flow_calculate_parameters}
\alias{general_radial_flow_calculate_parameters}
\title{general_radial_flow_calculate_parameters}
\usage{
general_radial_flow_calculate_parameters(ptest, par, hydraulic = TRUE)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{par}{A list with the slope and intercept (a and t0) of the straight line
fitted to the drawdown data using the Cooper-Jacob approach}

\item{hydraulic}{Logical flag to indicate if hydraulic parameters are calculated.
If False, the the statistcal parameter (a and t0) are calculated.}
}
\value{
A list with
\itemize{
\item Tr: Transmissivity
\item Ss: Storage coefficient
}
}
\description{
Calculates the hydraulic parameters form general_radial_flow model
}
\examples{
data("general_radial_flow1")
ptest.grf1 <- pumping_test("Well1", Q= 0.02322, r = 26.2, t = general_radial_flow1$t,
                           s = general_radial_flow1$s)
par <- list(rw = 0.1, rc = 0.1, rd = 1)
ptest.grf1$additional_parameters <- par
sol0 <-  general_radial_flow_solution_initial(ptest.grf1)
grf1.par <- general_radial_flow_calculate_parameters(ptest.grf1, sol0)
print(grf1.par)
}
\references{
Barker, J. A. A generalized radial flow model for hydraulic tests in fractured rock
Water Resources Research, 1988, 24, 1796-1804.
}
\seealso{
Other general_radial_flow functions: \code{\link{general_radial_flow_WF_LT}},
  \code{\link{general_radial_flow_solution_dlogt}},
  \code{\link{general_radial_flow_solution_initial}},
  \code{\link{general_radial_flow_solution}},
  \code{\link{general_radial_flow_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{general_radial_flow functions}
