% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_radial_flow_utilities.R
\name{general_radial_flow_solution_initial}
\alias{general_radial_flow_solution_initial}
\title{general_radial_flow_solution_initial}
\usage{
general_radial_flow_solution_initial(ptest)
}
\arguments{
\item{ptest}{A pumping_test object}
}
\value{
A list with
\itemize{
\item a: Slope of the straight line fitted to the drawdown data using the
Cooper-Jacob approach
\item t0: Intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach
\item n: Flow dimension (n=2)
}
}
\description{
This function calculate the initial parameters
}
\examples{
data("general_radial_flow1")
ptest.grf1 <- pumping_test("Well1", Q= 0.02322, r = 26.2, t = general_radial_flow1$t,
                           s = general_radial_flow1$s)
par <- list(rw = 0.1, rc = 0.1, rd = 1)
ptest.grf1$additional_parameters <- par
sol0 <-  general_radial_flow_solution_initial(ptest.grf1)
print(sol0)
}
\references{
Barker, J. A. A generalized radial flow model for hydraulic tests in fractured rock
Water Resources Research, 1988, 24, 1796-1804.
}
\seealso{
Other general_radial_flow functions: \code{\link{general_radial_flow_WF_LT}},
  \code{\link{general_radial_flow_calculate_parameters}},
  \code{\link{general_radial_flow_solution_dlogt}},
  \code{\link{general_radial_flow_solution}},
  \code{\link{general_radial_flow_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{general_radial_flow functions}
