% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_radial_flow_utilities.R
\name{general_radial_flow_well_function}
\alias{general_radial_flow_well_function}
\title{general_radial_flow_well_function}
\usage{
general_radial_flow_well_function(td, par)
}
\arguments{
\item{td}{Numeric vector with the dimensionless time}

\item{par}{A list with the parameters flow dimension(n), dimensionless radius (rd), and
the vector of stehfest coefficients (coeffs)}
}
\value{
A numeric vector with the dimensionless drawdown
}
\description{
Calculates the well function of the general radial flow model
}
\examples{
td <- logseq(-1, 4, 50)
par <- list(n = 2, rd = 10, coeffs = stehfest_coeffs(8))
W <- general_radial_flow_well_function(td, par)
plot(td, W, type = "l", log = "xy", main = "GRAL RADIAL FLOW: Well function")
}
\references{
Barker, J. A. A generalized radial flow model for hydraulic tests in fractured rock
Water Resources Research, 1988, 24, 1796-1804.
}
\seealso{
Other general_radial_flow functions: \code{\link{general_radial_flow_WF_LT}},
  \code{\link{general_radial_flow_calculate_parameters}},
  \code{\link{general_radial_flow_solution_dlogt}},
  \code{\link{general_radial_flow_solution_initial}},
  \code{\link{general_radial_flow_solution}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{general_radial_flow functions}
