% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_radial_flow_utilities.R
\name{general_radial_flow_WF_LT}
\alias{general_radial_flow_WF_LT}
\title{general_radial_flow_WF_LT}
\usage{
general_radial_flow_WF_LT(p, arg1, arg2)
}
\arguments{
\item{p}{Laplace Transform parameter}

\item{arg1}{Flow dimension (n)}

\item{arg2}{Dimensionless radius (rd)}
}
\value{
A numeric vector with the calculated drawdown in the Laplace domain
}
\description{
Function to calculate the general_radial_flow well function in the Laplace domain
}
\examples{
p <- 0.5
sd <- general_radial_flow_WF_LT(p, 2.0, 10)
print(sd)
}
\references{
Barker, J. A. A generalized radial flow model for hydraulic tests in fractured rock
Water Resources Research, 1988, 24, 1796-1804.
}
\seealso{
Other general_radial_flow functions: \code{\link{general_radial_flow_calculate_parameters}},
  \code{\link{general_radial_flow_solution_dlogt}},
  \code{\link{general_radial_flow_solution_initial}},
  \code{\link{general_radial_flow_solution}},
  \code{\link{general_radial_flow_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{general_radial_flow functions}
