% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest_data.R
\docType{data}
\name{gringarten}
\alias{gringarten}
\title{Single fracture in a confined aquifer. estimation using the Gringarten Solution.}
\format{A data frame with 26 rows and 2 variables
\describe{
\item{t}{time in seconds}
\item{s}{Drawdown in m}
}}
\source{
Gringarten A. C., Ramey H. J., & Raghavan R., 1975.
Applied Pressure Analysis for Fractured Wells, Petroleum Transactions,
AIME follows page 784.
}
\usage{
gringarten
}
\description{
The information of this pumping test is:
\itemize{
\item pumping rate (m3/s): Q <- 7.7e-4
\item Storage Coefficient S <- 1
\item Porosity : 0.12
}
}
\keyword{datasets}
