% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gringarten_utilities.R
\name{gringarten_calculate_parameters}
\alias{gringarten_calculate_parameters}
\title{gringarten_calculate_parameters}
\usage{
gringarten_calculate_parameters(ptest, par, hydraulic = TRUE)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{par}{A list with the values of slope and intercept (a and t0) of the straight
line fitted to the drawdown data using the Cooper-Jacob approach}

\item{hydraulic}{Logical flag to indicate if hydraulic parameters are calculated.
If False, the the statistcal parameter (a and t0) are calculated.}
}
\value{
A list with
\itemize{
\item Tr: Transmissivity
\item Sxf2: ????
}
}
\description{
Calculate the hydraulic parameters of the Gringarten solution
}
\references{
Gringarten A. C., Ramey H. J., & Raghavan R., 1975.
Applied Pressure Analysis for Fractured Wells, Petroleum Transactions,
AIME follows page 784.
}
\seealso{
Other gringarten functions: \code{\link{gringarten_solution_dlogt}},
  \code{\link{gringarten_solution_initial}},
  \code{\link{gringarten_solution}},
  \code{\link{gringarten_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{gringarten functions}
