% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gringarten_utilities.R
\name{gringarten_solution_dlogt}
\alias{gringarten_solution_dlogt}
\title{gringarten_solution_dlogt}
\usage{
gringarten_solution_dlogt(ptest, a, t0, t)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{a}{Slope  of the straight line fitted to the drawdown data using
the Cooper-Jacob approach}

\item{t0}{Intercept of the straight line fitted to the drawdown data using
the Cooper-Jacob approach}

\item{t}{A numeric vector with the time values}
}
\value{
A numeric vector with the derivative of the calculated drawdown
}
\description{
Function to calculate the derivative of drawdown with respect to logarithm of time using the
Gringarten Solution
}
\references{
Gringarten A. C., Ramey H. J., & Raghavan R., 1975.
Applied Pressure Analysis for Fractured Wells, Petroleum Transactions,
AIME follows page 784.
}
\seealso{
Other gringarten functions: \code{\link{gringarten_calculate_parameters}},
  \code{\link{gringarten_solution_initial}},
  \code{\link{gringarten_solution}},
  \code{\link{gringarten_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{gringarten functions}
