% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gringarten_utilities.R
\name{gringarten_solution_initial}
\alias{gringarten_solution_initial}
\title{gringarten_solution_initial}
\usage{
gringarten_solution_initial(ptest)
}
\arguments{
\item{ptest}{A pumping_test object}
}
\value{
A list with
\itemize{
\item a: Slope of the straight line fitted to the drawdown data using the
Cooper-Jacob approach
\item t0: Intercept  of the straight line fitted to the drawdown data using the
Cooper-Jacob approach
}
}
\description{
Function to calculate the initial parameters of the Gringarten Solution
}
\references{
Gringarten A. C., Ramey H. J., & Raghavan R., 1975.
Applied Pressure Analysis for Fractured Wells, Petroleum Transactions,
AIME follows page 784.
}
\seealso{
Other gringarten functions: \code{\link{gringarten_calculate_parameters}},
  \code{\link{gringarten_solution_dlogt}},
  \code{\link{gringarten_solution}},
  \code{\link{gringarten_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{gringarten functions}
