% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gringarten_utilities.R
\name{gringarten_well_function}
\alias{gringarten_well_function}
\title{gringarten_well_function}
\usage{
gringarten_well_function(td, par)
}
\arguments{
\item{td}{A numeric vector with the dimensionless time}

\item{par}{A list with the values of slope and intercept (a and t0) of the straight
line fitted to the drawdown data using the Cooper-Jacob approach}
}
\value{
A numeric vector with the calculated drawdown
}
\description{
Function to calculate the drawdown using the Gringarten Solution
}
\references{
Gringarten A. C., Ramey H. J., & Raghavan R., 1975.
Applied Pressure Analysis for Fractured Wells, Petroleum Transactions,
AIME follows page 784.
}
\seealso{
Other gringarten functions: \code{\link{gringarten_calculate_parameters}},
  \code{\link{gringarten_solution_dlogt}},
  \code{\link{gringarten_solution_initial}},
  \code{\link{gringarten_solution}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{gringarten functions}
