% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest_data.R
\docType{data}
\name{hantush_jacob}
\alias{hantush_jacob}
\title{estimation of a pumping test in a Confined aquifer with leakage using the
Hantush-Jacob solution}
\format{A data frame with 43 rows and 2 variables:
\describe{
  \item{t}{time, in seconds}
  \item{s}{drawdown, in meters}
}}
\source{
Hall, P. Water well and aquifer test analysis 1996, Water Resources
Publications, LLC, 412 pp. Data from Fig 11.14 p.267-268
}
\usage{
hantush_jacob
}
\description{
The information for this pumping test is:
\itemize{
\item Pumping rate (in m3/s):  Q <- 6.309e-3
\item Distance to the pumping well (in m): r <- 3.048
\item Thickness of the aquitard (in m): B <- 6.096
}
}
\keyword{datasets}
