% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hantush_jacob_utilities.R
\name{hantush_jacob_calculate_parameters}
\alias{hantush_jacob_calculate_parameters}
\title{hantush_jacob_calculate_parameters}
\usage{
hantush_jacob_calculate_parameters(ptest, par, hydraulic = TRUE)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{par}{A list with the values of a (slope) and t0 (Intercept) of the straight line fitted
to the drawdown data using the Cooper-Jacob approach,  and lambda is the leakage factor.}

\item{hydraulic}{Logical flag to indicate if hydraulic parameters are calculated. If False,
the the statistcal parameter are calculated.}
}
\value{
A list with
\itemize{
\item Tr = Transmissivity
\item Ss = Storage coefficient
\item Ka = Aquitard Hydraulic Conductivity
\item radius_influence
}
}
\description{
Function to calculate the hydraulic parameters of the Hantush-Jacob model
}
\examples{
data(hantush_jacob)
ptest <- pumping_test('Well1', Q = 6.309e-3, r = 3.048, t=hantush_jacob$t, s = hantush_jacob$s)
ptest$additional_parameters$B <- 6.096
sol0 <- hantush_jacob_solution_initial(ptest)
hantush.par <- hantush_jacob_calculate_parameters(ptest,sol0)
print(hantush.par)
}
\references{
Hantush, M. S. Flow to wells in aquifers separated by a semipervious layer. Journal
of Geophysical Research, 1967, 72, 1709.
}
\seealso{
Other hantush_jacob functions: \code{\link{hantush_jacob_WF_LT_dlogt}},
  \code{\link{hantush_jacob_WF_LT}},
  \code{\link{hantush_jacob_solution_dlogt}},
  \code{\link{hantush_jacob_solution_initial}},
  \code{\link{hantush_jacob_solution}},
  \code{\link{hantush_jacob_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{hantush_jacob functions}
