% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hantush_jacob_utilities.R
\name{hantush_jacob_solution_dlogt}
\alias{hantush_jacob_solution_dlogt}
\title{hantush_jacob_solution_logdt}
\usage{
hantush_jacob_solution_dlogt(ptest, a, t0, lambda, t)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{a}{Slope of the straight line fitted to the drawdown data using the
Cooper-Jacob approach}

\item{t0}{Intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach}

\item{lambda}{Leakage factor defined as
\deqn{\lambda = \sqrt{\frac{b^{'}T}{k_{h}^{'}}}}
where \eqn{b^{'}} and \eqn{k_{h}^{'}} are the thickness and the vertical hydraulic
conductivity of the aquitard, and \eqn{T} is the transmissivity of the underlying
aquifer.}

\item{t}{Numeric vector with the time values}
}
\value{
This function returns a list with the x and y values of the derivative of drawdown with
respect to log of time
}
\description{
Function to calculate the derivative of drawdown with respect to the log of time
}
\examples{
data(hantush_jacob)
ptest <- pumping_test('Well1', Q = 6.309e-3, r = 3.048, t=hantush_jacob$t, s = hantush_jacob$s)
ptest$additional_parameters$B <- 6.096
sol0 <- hantush_jacob_solution_initial(ptest)
sol_dlogt <- hantush_jacob_solution_dlogt(ptest, sol0$a, sol0$t0, sol0$lambda, ptest$t)
print(sol_dlogt)
}
\references{
Hantush, M. S. Flow to wells in aquifers separated by a semipervious layer. Journal
of Geophysical Research, 1967, 72, 1709.
}
\seealso{
Other hantush_jacob functions: \code{\link{hantush_jacob_WF_LT_dlogt}},
  \code{\link{hantush_jacob_WF_LT}},
  \code{\link{hantush_jacob_calculate_parameters}},
  \code{\link{hantush_jacob_solution_initial}},
  \code{\link{hantush_jacob_solution}},
  \code{\link{hantush_jacob_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{hantush_jacob functions}
