% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hantush_jacob_utilities.R
\name{hantush_jacob_solution_initial}
\alias{hantush_jacob_solution_initial}
\title{hantush_jacob_solutionb_initial}
\usage{
hantush_jacob_solution_initial(ptest)
}
\arguments{
\item{ptest}{A pumping_test object}
}
\value{
A list with
\itemize{
\item a: Slope of the straight line fitted to the drawdown data using the
Cooper-Jacob approach
\item t0: Intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach
\item lambda: Leakage factor defined as
\deqn{\lambda = \sqrt{\frac{b^{'}T}{k_{h}^{'}}} }
}
where \eqn{b^{'}} and \eqn{k_{h}^{'}} are the thickness and the vertical hydraulic
conductivity of the aquitard, and \eqn{T} is the transmissivity of the underlying
aquifer.
}
\description{
Function to calculate the initial parameters of the Hantush-Jacob model
}
\examples{
data(hantush_jacob)
ptest <- pumping_test('Well1', Q = 6.309e-3, r = 3.048, t=hantush_jacob$t, s = hantush_jacob$s)
sol0 <- hantush_jacob_solution_initial(ptest)
print(sol0)
}
\references{
Hantush, M. S. Flow to wells in aquifers separated by a semipervious layer. Journal
of Geophysical Research, 1967, 72, 1709.
}
\seealso{
Other hantush_jacob functions: \code{\link{hantush_jacob_WF_LT_dlogt}},
  \code{\link{hantush_jacob_WF_LT}},
  \code{\link{hantush_jacob_calculate_parameters}},
  \code{\link{hantush_jacob_solution_dlogt}},
  \code{\link{hantush_jacob_solution}},
  \code{\link{hantush_jacob_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{hantush_jacob functions}
