% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hantush_jacob_utilities.R
\name{hantush_jacob_well_function}
\alias{hantush_jacob_well_function}
\title{hantush_jacob_well_function}
\usage{
hantush_jacob_well_function(td, par)
}
\arguments{
\item{td}{Dimensionless time}

\item{par}{List with the beta parameter and the coefficients used in the Stehfest
inversion approach. The beta parameter is defined as
\deqn{\beta = \frac{r}{4} \sqrt{\frac{k_{h}^{'}S_{s}^{'}}{TS}}}
where \eqn{T} and \eqn{S} are the transmissivity and storage coefficient of the
aquifer, and \eqn{k_{h}^{'}} and \eqn{S_{s}^{'}} are the vertical hydraulic
conductivity and the Storage coefficient of the aquitard respectively.}
}
\value{
The value of the dimensionless drawdown of the Hantush-Jacob model
}
\description{
Well function of the Hantush-Jacob model
}
\examples{
td <- logseq(-2, 4, 50)
par <- list( beta = 0.5)
res <- hantush_jacob_well_function(td, par)
print(res)
}
\references{
Hantush, M. S. Flow to wells in aquifers separated by a semipervious layer. Journal
of Geophysical Research, 1967, 72, 1709.
}
\seealso{
Other hantush_jacob functions: \code{\link{hantush_jacob_WF_LT_dlogt}},
  \code{\link{hantush_jacob_WF_LT}},
  \code{\link{hantush_jacob_calculate_parameters}},
  \code{\link{hantush_jacob_solution_dlogt}},
  \code{\link{hantush_jacob_solution_initial}},
  \code{\link{hantush_jacob_solution}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{hantush_jacob functions}
