% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hantush_jacob_utilities.R
\name{hantush_jacob_WF_LT_dlogt}
\alias{hantush_jacob_WF_LT_dlogt}
\title{hantush_jacob_WF_LT_dlogt}
\usage{
hantush_jacob_WF_LT_dlogt(p, arg1, ...)
}
\arguments{
\item{p}{Laplace parameter}

\item{arg1}{Value of beta}

\item{...}{Additional parameters}
}
\value{
Numeric vector with the values of the derivative of drawdown with respect to logarithm of
time for the Hantush-Jacob solution in the Laplace domain
}
\description{
Calculates the values of the derivative of drawdown with respect to logarithm of
time for the Hantush-Jacob solution in the Laplace domain
}
\examples{
p <- 0.5
beta <- 0.5
res <- hantush_jacob_WF_LT_dlogt(p, beta)
print(res)
}
\references{
Hantush, M. S. Flow to wells in aquifers separated by a semipervious layer. Journal
of Geophysical Research, 1967, 72, 1709.
}
\seealso{
Other hantush_jacob functions: \code{\link{hantush_jacob_WF_LT}},
  \code{\link{hantush_jacob_calculate_parameters}},
  \code{\link{hantush_jacob_solution_dlogt}},
  \code{\link{hantush_jacob_solution_initial}},
  \code{\link{hantush_jacob_solution}},
  \code{\link{hantush_jacob_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{hantush_jacob functions}
