% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest_data.R
\docType{data}
\name{hazel}
\alias{hazel}
\title{estimation of Step-drawdown test with Eden and Hazel (1973) method.}
\format{A data frame with 175 rows and 2 variables
\describe{
\item{t}{time in seconds}
\item{s}{Drawdown in m}
}}
\source{
Clark, L. (1977) The analysis and planning of step-drawdown tests.
Quarterly Journal of Engineering Geology and Hydrogeology 10(2):125-143
}
\usage{
hazel
}
\description{
The pumping rates in m3/s are given by:
\itemize{
\item pumping rates:  qp <- c(0.01511574,0.01959491,0.02804398,0.03774306,0.04738426,0.05809028)
\item final time for each step: tp <- c(10800, 21600, 32400, 43200, 54000, 64800)
}
}
\keyword{datasets}
