% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvorslev_utilities.R
\name{hvorslev_calculate_parameters}
\alias{hvorslev_calculate_parameters}
\title{hvorslev_calculate_parameters}
\usage{
hvorslev_calculate_parameters(ptest, par, hydraulic = TRUE)
}
\arguments{
\item{ptest}{A pumpingtest object}

\item{par}{A list with the fit parameters}

\item{hydraulic}{A logical flag indicating if hydraulic parameters are required.}
}
\value{
A list with the time t37.
}
\description{
Function calculate the drawdown using the Hvorslev Solution during a slug test.
}
\references{
Hvorslev, M.J., 1951. Time Lag and Soil Permeability in Ground-Water Observations,
Bull. No. 36, Waterways Exper. Sta. Corps of Engrs, U.S. Army, Vicksburg, Mississippi,
pp. 1-50.
}
\seealso{
Other hvorslev functions: \code{\link{hvorslev_initial_solution}},
  \code{\link{hvorslev_solution}},
  \code{\link{hvorslev_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{hvorslev functions}
