% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvorslev_utilities.R
\name{hvorslev_solution}
\alias{hvorslev_solution}
\title{hvorslev_solution}
\usage{
hvorslev_solution(ptest, t0, t)
}
\arguments{
\item{ptest}{A pumpingtest object}

\item{t0}{Time for which the normalized drawdown is equal to 0.37}

\item{t}{A numeric vector with the time}
}
\value{
A numeric vector with the calculated drawdown
}
\description{
Function calculate the drawdown using the Hvorslev Solution during a slug test.
}
\references{
Hvorslev, M.J., 1951. Time Lag and Soil Permeability in Ground-Water Observations,
Bull. No. 36, Waterways Exper. Sta. Corps of Engrs, U.S. Army, Vicksburg, Mississippi,
pp. 1-50.
}
\seealso{
Other hvorslev functions: \code{\link{hvorslev_calculate_parameters}},
  \code{\link{hvorslev_initial_solution}},
  \code{\link{hvorslev_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{hvorslev functions}
