% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvorslev_utilities.R
\name{hvorslev_well_function}
\alias{hvorslev_well_function}
\title{hvorslev_well_function}
\usage{
hvorslev_well_function(td, par)
}
\arguments{
\item{td}{A numeric vector with the dimensionless time}

\item{par}{A list with the values of intercept t0 of the straight line fitted to the
drawdown data.}
}
\value{
A numeric vector with the calculated drawdown
}
\description{
Function calculate the drawdown using the Hvorslev Solution during a slug test.
}
\references{
Hvorslev, M.J., 1951. Time Lag and Soil Permeability in Ground-Water Observations,
Bull. No. 36, Waterways Exper. Sta. Corps of Engrs, U.S. Army, Vicksburg, Mississippi,
pp. 1-50.
}
\seealso{
Other hvorslev functions: \code{\link{hvorslev_calculate_parameters}},
  \code{\link{hvorslev_initial_solution}},
  \code{\link{hvorslev_solution}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{hvorslev functions}
