% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest.R
\name{hydraulic.parameter.names<-}
\alias{hydraulic.parameter.names<-}
\title{hydraulic.parameter.names <-}
\usage{
hydraulic.parameter.names(x) <- value
}
\arguments{
\item{x}{A pumping_test object}

\item{value}{A list with the names of the hydraulic parameters}
}
\value{
The pumping_test object with the names of the hydraulic parameters
}
\description{
Function to assign name of the hydraulic parameters to a  pumping_test object. This is
useful in plotting results
}
\seealso{
Other base functions: \code{\link{additional.parameters<-}},
  \code{\link{confint.pumping_test}},
  \code{\link{confint_bootstrap}},
  \code{\link{confint_jackniffe}},
  \code{\link{confint_wald}}, \code{\link{estimated<-}},
  \code{\link{evaluate}}, \code{\link{fit.optimization}},
  \code{\link{fit.parameters<-}},
  \code{\link{fit.sampling}}, \code{\link{fit}},
  \code{\link{hydraulic.parameters<-}},
  \code{\link{model.parameters}}, \code{\link{model<-}},
  \code{\link{plot.pumping_test}},
  \code{\link{plot_model_diagnostic}},
  \code{\link{print.pumping_test}},
  \code{\link{pumping_test}}, \code{\link{simulate}},
  \code{\link{summary.pumping_test}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{base functions}
