% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest_data.R
\docType{data}
\name{jacob_lohman}
\alias{jacob_lohman}
\title{Constant head test}
\format{A data frame with 19 rows and 2 variables:
\describe{
  \item{t}{time, in seconds}
  \item{q}{discharge, in meters3/second}
}}
\source{
Lohman (1965) Geology and artesian water supply of the Grand Junction area,
Colorado: U.S. Geol. Survey Prof. Paper 451, 149p. Tables 6 and 7, well 28
}
\usage{
jacob_lohman
}
\description{
The information for this pumping test is:
\itemize{
\item Discharge (in m3/s): Q <- c(28.142,0.084)
\item well radius (in m): r <- 0.084
}
}
\keyword{datasets}
