% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacob_lohman_utilities.R
\name{jacob_lohman_F_calculate_parameters}
\alias{jacob_lohman_F_calculate_parameters}
\title{jacob_lohman_F_calculate_parameters}
\usage{
jacob_lohman_F_calculate_parameters(ptest, par, hydraulic = TRUE)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{par}{A list with the Slope (a) and the Intercept (t0) of the straight line fitted to
the drawdown data using the Cooper-Jacob approach}

\item{hydraulic}{Logical flag to indicate if hydraulic parameters are calculated. If False,
the the statistcal parameter (a and t0) are calculated.}
}
\value{
A list with
\itemize{
\item Tr: Transmisivity.
\item Ss: The Storage coefficient
\item radius_influence
}
}
\description{
Calculate hydraulic parameters of the Jacob-Lohman Solution
}
\examples{
data(jacob_lohman)
s <- vector("numeric", length(jacob_lohman$t))
s[1:length(jacob_lohman$t)] <- 0
ptest <- pumping_test("Well1", Q = jacob_lohman$q, r = 0.84, t = jacob_lohman$t, s = s)
ptest$additional_parameters <- list(rw = 0.084, s0 = 28.142)
sol0 <- jacob_lohman_F_solution_initial(ptest)
jl.par <- jacob_lohman_F_calculate_parameters(ptest, sol0)
print(jl.par)
}
\references{
Jacob, C. E. & Lohman, S. W. Nonsteady Flow to a Well Of Constant Drawdown. American
Geophysical Union, 1952, 33, 10.
}
\seealso{
Other jacob_lohman functions: \code{\link{jacob_lohman_F_solution_initial}},
  \code{\link{jacob_lohman_F_solution}},
  \code{\link{jacob_lohman_F_well_function}},
  \code{\link{jacob_lohman_G_WF_LT}},
  \code{\link{jacob_lohman_G_calculate_parameters}},
  \code{\link{jacob_lohman_G_solution_initial}},
  \code{\link{jacob_lohman_G_solution}},
  \code{\link{jacob_lohman_G_well_function}},
  \code{\link{jacob_lohman_WF_F_LT}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{jacob_lohman functions}
