% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacob_lohman_utilities.R
\name{jacob_lohman_F_solution}
\alias{jacob_lohman_F_solution}
\title{jacob_lohman_F_solution}
\usage{
jacob_lohman_F_solution(ptest, a, t0, t)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{a}{Slope of the straight line fitted to the drawdown data using the
Cooper-Jacob approach}

\item{t0}{Intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach}

\item{t}{Numeric vector with the times at which measurements were taken}
}
\value{
A numeric vector with the values of the drawdown
}
\description{
Function to calculate the drawdown in a constant rate test using the Jacob-Lohman Solution
}
\examples{
data(jacob_lohman)
s <- vector("numeric", length(jacob_lohman$t))
s[1:length(jacob_lohman$t)] <- 0
ptest <- pumping_test("Well1", Q = jacob_lohman$q, r = 0.84, t = jacob_lohman$t, s = s)
ptest$additional_parameters <- list(rw = 0.084, s0 = 28.142)
sol0 <- jacob_lohman_F_solution_initial(ptest)
sol <- jacob_lohman_F_solution(ptest, sol0$a, sol0$t0, ptest$t)
print(sol)
}
\references{
Jacob, C. E. & Lohman, S. W. Nonsteady Flow to a Well Of Constant Drawdown. American
Geophysical Union, 1952, 33, 10.
}
\seealso{
Other jacob_lohman functions: \code{\link{jacob_lohman_F_calculate_parameters}},
  \code{\link{jacob_lohman_F_solution_initial}},
  \code{\link{jacob_lohman_F_well_function}},
  \code{\link{jacob_lohman_G_WF_LT}},
  \code{\link{jacob_lohman_G_calculate_parameters}},
  \code{\link{jacob_lohman_G_solution_initial}},
  \code{\link{jacob_lohman_G_solution}},
  \code{\link{jacob_lohman_G_well_function}},
  \code{\link{jacob_lohman_WF_F_LT}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{jacob_lohman functions}
