% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacob_lohman_utilities.R
\name{jacob_lohman_F_well_function}
\alias{jacob_lohman_F_well_function}
\title{jacob_lohman_F_well_function}
\usage{
jacob_lohman_F_well_function(td, par)
}
\arguments{
\item{td}{Numeric vector with the Dimensionless time}

\item{par}{A list with the parameters}
}
\value{
A numeric vector with the dimensionless drawdowns
}
\description{
This calculates the well function of the Jacob-Lohman solution for a constant head aquifer
(artesian aquifer). This function calculates the drawdown.
}
\examples{
td <- logseq(2, 6, 100)
par <- list(rho = 5, coeffs = stehfest_coeffs(8))
sd <- jacob_lohman_F_well_function(td, par)
}
\references{
Jacob, C. E. & Lohman, S. W. Nonsteady Flow to a Well Of Constant Drawdown. American
Geophysical Union, 1952, 33, 10.
}
\seealso{
Other jacob_lohman functions: \code{\link{jacob_lohman_F_calculate_parameters}},
  \code{\link{jacob_lohman_F_solution_initial}},
  \code{\link{jacob_lohman_F_solution}},
  \code{\link{jacob_lohman_G_WF_LT}},
  \code{\link{jacob_lohman_G_calculate_parameters}},
  \code{\link{jacob_lohman_G_solution_initial}},
  \code{\link{jacob_lohman_G_solution}},
  \code{\link{jacob_lohman_G_well_function}},
  \code{\link{jacob_lohman_WF_F_LT}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{jacob_lohman functions}
