% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacob_lohman_utilities.R
\name{jacob_lohman_G_calculate_parameters}
\alias{jacob_lohman_G_calculate_parameters}
\title{jacob_lohman_G_calculate_parameters}
\usage{
jacob_lohman_G_calculate_parameters(ptest, par)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{par}{A list with the slope and the intercept (a and t0) of the straight line fitted to
the drawdown data using the Cooper-Jacob approach}
}
\value{
A list with
\itemize{
\item Tr Transmissivity
\item Ss Storage coefficient
\item radius_influence
}
}
\description{
Function to calculate the hydraulic parameters for the Jacob-Lohman solution (Discharge)
}
\examples{
data(jacob_lohman)
s <- vector("numeric", length(jacob_lohman$t))
s[1:length(jacob_lohman$t)] <- 0
ptest <- pumping_test("Well1", Q = jacob_lohman$q, r = 0.0, t = jacob_lohman$t, s = s)
sol0 <- jacob_lohman_G_solution_initial(ptest)
jl.par <- jacob_lohman_G_calculate_parameters(ptest, sol0)
print(jl.par)
}
\references{
Jacob, C. E. & Lohman, S. W. Nonsteady Flow to a Well Of Constant Drawdown. American
Geophysical Union, 1952, 33, 10.
}
\seealso{
Other jacob_lohman functions: \code{\link{jacob_lohman_F_calculate_parameters}},
  \code{\link{jacob_lohman_F_solution_initial}},
  \code{\link{jacob_lohman_F_solution}},
  \code{\link{jacob_lohman_F_well_function}},
  \code{\link{jacob_lohman_G_WF_LT}},
  \code{\link{jacob_lohman_G_solution_initial}},
  \code{\link{jacob_lohman_G_solution}},
  \code{\link{jacob_lohman_G_well_function}},
  \code{\link{jacob_lohman_WF_F_LT}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{jacob_lohman functions}
