% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{log_derivative}
\alias{log_derivative}
\title{log_derivative}
\usage{
log_derivative(t, s, d = 2, method = "central", return.pos = T)
}
\arguments{
\item{t}{vector with time values}

\item{s}{vector with drawdown values}

\item{d}{Derivative parameter. if method equals to bourdet then d is equal to the number of
adjacent values used in the derivative calculation. If method is equal to spline then d is equal
to the number of knots used in the interpolation of the drawdown data. In this case a value of d=20 to
d=30 is recommended. If method is equal to spane then d is equal to the number of points used in the
linear regression approach.}

\item{method}{Method to calculate the derivative (central, horner, bourdet and
spline)}

\item{return.pos}{Logical flag to return only the positive values of the log-derivative}
}
\value{
This function returns a list with components named as x and y that contains the
log_derivative y evaluated at specific points x.
}
\description{
Function to calculate the derivative of drawdown wrt the log of time
}
\examples{
# Load test data
data(theis)
# calculate derrivative of s wrt to logt using central differences
logd.central <- log_derivative(theis$t, theis$s)
# calculate derrivative of s wrt to logt using Horner method
logd.horner <- log_derivative(theis$t, theis$s, method = "horner")
# calculate derrivative of s wrt to logt using Bourdet method
logd.bourdet <- log_derivative(theis$t, theis$s, method = "bourdet")
# plot results
plot(logd.central$x, logd.central$y, type = "p")
points(logd.horner$x, logd.horner$y, col = "red")
points(logd.bourdet$x, logd.bourdet$y, col = "blue")
}
\seealso{
Other log_derivative functions: \code{\link{log_derivative_bourdet}},
  \code{\link{log_derivative_central}},
  \code{\link{log_derivative_horner}},
  \code{\link{log_derivative_kernelreg}},
  \code{\link{log_derivative_locpol}},
  \code{\link{log_derivative_lokern}},
  \code{\link{log_derivative_lpridge}},
  \code{\link{log_derivative_smoothspline}},
  \code{\link{log_derivative_spane}},
  \code{\link{log_derivative_spline}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{log_derivative functions}
