% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{log_derivative_bourdet}
\alias{log_derivative_bourdet}
\title{log_derivative_bourdet}
\usage{
log_derivative_bourdet(t, s, d = 2, return.pos = T, log = T)
}
\arguments{
\item{t}{Numeric vector with the time}

\item{s}{Numeric vector with the drawdown}

\item{d}{Numeric value}

\item{return.pos}{Logical flag to return only the positive values of the log-derivative (default = TRUE)}

\item{log}{Logical flag to indicate that natural logarithm (a log to the base e) is used in the derivative 
calculation (default = TRUE). Logarithm to the base 10 is used if FALSE.}
}
\value{
A list with
\itemize{
  \item x: Numeric vector with the x coordinates where the log-derivative is evaluated
  \item y: Numeric vector with the value of the log-derivative
}
}
\description{
Function to calculate the derivative of the drawdown with respect to the derivative
of log time using the approach proposed by Bourdet
}
\examples{
data(boulton)
t <- boulton$t
s <- boulton$s
ptest <- pumping_test('Well1', Q = 0.03, r = 20, t = t, s = s)
#
dptest.bd <- log_derivative_bourdet(ptest$t, ptest$s, d = 10)
plot(t, s, type= "p", log = "xy", ylim = c(1e-3,2))
points(dptest.bd$x, dptest.bd$y, col = "red")
}
\references{
Bourdet, D.; Whittle, T.; Douglas, A. & Pirard, Y. A new set of type curves
 simplifies well test analysis World Oil, 1983.
}
\seealso{
Other log_derivative functions: \code{\link{log_derivative_central}},
  \code{\link{log_derivative_horner}},
  \code{\link{log_derivative_kernelreg}},
  \code{\link{log_derivative_locpol}},
  \code{\link{log_derivative_lokern}},
  \code{\link{log_derivative_lpridge}},
  \code{\link{log_derivative_smoothspline}},
  \code{\link{log_derivative_spane}},
  \code{\link{log_derivative_spline}},
  \code{\link{log_derivative}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{log_derivative functions}
