% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{log_derivative_lokern}
\alias{log_derivative_lokern}
\title{log_derivative_lokern}
\usage{
log_derivative_lokern(t, s)
}
\arguments{
\item{t}{Numeric vector with the time}

\item{s}{Numeric vector with the drawdown}
}
\value{
A list with
\itemize{
  \item x: Numeric vector with the x coordinates where the log-derivative is evaluated
  \item y: Numeric vector with the value of the log-derivative
}
}
\description{
Function to calculate the derivative of the drawdown with respect to the derivative
of log time using local kernel regression of the measured data.
}
\examples{
data(boulton)
t <- boulton$t
s <- boulton$s
ptest <- pumping_test('Well1', Q = 0.03, r = 20, t = t, s = s)
#
dptest.lp <- log_derivative_lokern(ptest$t, ptest$s)
plot(t, s, type = "p", log = "xy", ylim = c(1e-3,2))
points(dptest.lp$x, dptest.lp$y, col = "red")
}
\seealso{
Other log_derivative functions: \code{\link{log_derivative_bourdet}},
  \code{\link{log_derivative_central}},
  \code{\link{log_derivative_horner}},
  \code{\link{log_derivative_kernelreg}},
  \code{\link{log_derivative_locpol}},
  \code{\link{log_derivative_lpridge}},
  \code{\link{log_derivative_smoothspline}},
  \code{\link{log_derivative_spane}},
  \code{\link{log_derivative_spline}},
  \code{\link{log_derivative}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{log_derivative functions}
