% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{log_derivative_spline}
\alias{log_derivative_spline}
\title{log_derivative_spline}
\usage{
log_derivative_spline(t, s, n = 20)
}
\arguments{
\item{t}{Numeric vector with the time}

\item{s}{Numeric vector with the drawdown}

\item{n}{Number of points where the derivative is calculated}
}
\value{
A list with
\itemize{
  \item x: Numeric vector with the x coordinates where the log-derivative is evaluated
  \item y: Numeric vector with the value of the log-derivative
}
}
\description{
Function to calculate the derivative of the drawdown with respect to the derivative
of log time using the interpolation of the measured points with spline
}
\examples{
# Load test data
data(boulton)
t <- boulton$t
s <- boulton$s
ptest <- pumping_test('Well1', Q = 0.03, r = 20, t = t, s = s)
#
dptest.sp <- log_derivative_spline(ptest$t, ptest$s, n = 30)
plot(t, s, type="p", log="xy", ylim = c(1e-3,2))
points(dptest.sp$x, dptest.sp$y, col = "red")
}
\references{
Renard, P.; Glenz, D. , Mejias, M. Understanding diagnostic plots for well-test estimation, Hydrogeology Journal, 2008, 17, 589-600
}
\seealso{
Other log_derivative functions: \code{\link{log_derivative_bourdet}},
  \code{\link{log_derivative_central}},
  \code{\link{log_derivative_horner}},
  \code{\link{log_derivative_kernelreg}},
  \code{\link{log_derivative_locpol}},
  \code{\link{log_derivative_lokern}},
  \code{\link{log_derivative_lpridge}},
  \code{\link{log_derivative_smoothspline}},
  \code{\link{log_derivative_spane}},
  \code{\link{log_derivative}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{log_derivative functions}
