% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_function_utilities.R
\name{loglikelihood_bc}
\alias{loglikelihood_bc}
\title{loglikelihood_bc}
\usage{
loglikelihood_bc(par, ptest, model)
}
\arguments{
\item{par}{A list with the parameters of the model.}

\item{ptest}{A pumping_test object.}

\item{model}{A character string with the name of the model.}
}
\value{
The numeric value of the loglikelihood function.
}
\description{
Function to calculate the loglikelihood function of the pumping test data using the Box-Cox
transform assuming that the residuals between the measured and simulated drawdown follow
a normal distribution.
}
\seealso{
Other objective_function functions: \code{\link{box_cox_transf}},
  \code{\link{loglikelihood}},
  \code{\link{max_absolute_deviation}},
  \code{\link{mean_absolute_deviation}},
  \code{\link{mean_square_error}},
  \code{\link{residual_sum_squares}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{objective_function functions}
