% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{logseq}
\alias{logseq}
\title{logseq}
\usage{
logseq(from = 1, to = 1, n)
}
\arguments{
\item{from, to}{log10 of initial and final points in the sequence}

\item{n}{step}
}
\value{
Sequence with log increments
}
\description{
Function to generate a sequence with log increments
}
\examples{
s <- logseq(0, 2, 10)
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{utility functions}
