% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_function_utilities.R
\name{mean_absolute_deviation}
\alias{mean_absolute_deviation}
\title{mean_absolute_deviation}
\usage{
mean_absolute_deviation(par, ptest, model)
}
\arguments{
\item{par}{A list with the parameters of the model.}

\item{ptest}{A pumping_test object.}

\item{model}{A character string with the name of the model.}
}
\value{
The numeric value of the mean absolute deviation.
}
\description{
Function to calculate the mean absolute deviation between the measured and simulated drawdown.
}
\seealso{
Other objective_function functions: \code{\link{box_cox_transf}},
  \code{\link{loglikelihood_bc}},
  \code{\link{loglikelihood}},
  \code{\link{max_absolute_deviation}},
  \code{\link{mean_square_error}},
  \code{\link{residual_sum_squares}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{objective_function functions}
