% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_function_utilities.R
\name{mean_square_error}
\alias{mean_square_error}
\title{mean_square_error}
\usage{
mean_square_error(par, ptest, model)
}
\arguments{
\item{par}{A list with the parameters of the model.}

\item{ptest}{A pumping_test object.}

\item{model}{A character string with the name of the model.}
}
\value{
The numeric value of the mean squared error
}
\description{
This function calculates the mean square error between the measured and calculated
drawdown
}
\examples{
data(theis)
}
\seealso{
Other objective_function functions: \code{\link{box_cox_transf}},
  \code{\link{loglikelihood_bc}},
  \code{\link{loglikelihood}},
  \code{\link{max_absolute_deviation}},
  \code{\link{mean_absolute_deviation}},
  \code{\link{residual_sum_squares}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{objective_function functions}
