% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest.R
\name{model<-}
\alias{model<-}
\title{model<-}
\usage{
model(x) <- value
}
\arguments{
\item{x}{A pumping_test object}

\item{value}{Character sting wit the model used in the intepretation}
}
\value{
The pumping_test object with the model assigned
}
\description{
Function to assign the model for estimation to a  pumping_test object
}
\examples{
data(theis)
ptest.theis <- pumping_test('Well1', Q = 1.38e-2, r = 250, t = theis$t, s = theis$s)
ptest.theis.fit <- fit(ptest.theis, 'theis')
hydraulic.parameters(ptest.theis) <- ptest.theis.fit$hydraulic_parameters
model(ptest.theis) <- 'theis'
}
\seealso{
Other base functions: \code{\link{additional.parameters<-}},
  \code{\link{confint.pumping_test}},
  \code{\link{confint_bootstrap}},
  \code{\link{confint_jackniffe}},
  \code{\link{confint_wald}}, \code{\link{estimated<-}},
  \code{\link{evaluate}}, \code{\link{fit.optimization}},
  \code{\link{fit.parameters<-}},
  \code{\link{fit.sampling}}, \code{\link{fit}},
  \code{\link{hydraulic.parameter.names<-}},
  \code{\link{hydraulic.parameters<-}},
  \code{\link{model.parameters}},
  \code{\link{plot.pumping_test}},
  \code{\link{plot_model_diagnostic}},
  \code{\link{print.pumping_test}},
  \code{\link{pumping_test}}, \code{\link{simulate}},
  \code{\link{summary.pumping_test}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{base functions}
