% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest.R
\name{model.parameters}
\alias{model.parameters}
\title{model.parameters}
\usage{
model.parameters(model, hydraulic = TRUE)
}
\arguments{
\item{model}{A character string specifying the analytical model.}

\item{hydraulic}{A logical flag indicating the type of parameters to return (TRUE for
hydraulic, FALSE for fit parameters)}
}
\value{
A list with the corresponding hydraulic or fit parameters.
}
\description{
This function returns a list with the hydraulic or fit parameters of a specific analytical
solution.
}
\examples{
model <- 'theis'
# Hydraulic parameters of the Theis model
print(model.parameters(model))
# Fit parameters of the Theis model
print(model.parameters(model, FALSE))
}
\seealso{
Other base functions: \code{\link{additional.parameters<-}},
  \code{\link{confint.pumping_test}},
  \code{\link{confint_bootstrap}},
  \code{\link{confint_jackniffe}},
  \code{\link{confint_wald}}, \code{\link{estimated<-}},
  \code{\link{evaluate}}, \code{\link{fit.optimization}},
  \code{\link{fit.parameters<-}},
  \code{\link{fit.sampling}}, \code{\link{fit}},
  \code{\link{hydraulic.parameter.names<-}},
  \code{\link{hydraulic.parameters<-}},
  \code{\link{model<-}}, \code{\link{plot.pumping_test}},
  \code{\link{plot_model_diagnostic}},
  \code{\link{plot_sample_influence}},
  \code{\link{plot_uncert}},
  \code{\link{print.pumping_test}},
  \code{\link{pumping_test}}, \code{\link{simulate}},
  \code{\link{summary.pumping_test}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{base functions}
