% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest_data.R
\docType{data}
\name{neuzil}
\alias{neuzil}
\title{estimation of the Shut-in slug test (pulse) using Neuzil (1982) solution}
\format{A data frame with 20 rows and 2 variables:
\describe{
  \item{t}{time, in seconds}
  \item{s}{drawdown, in meters}
}}
\source{
Batu, V., Aquifer Hydraulics: A Comprehensive Guide to Hydrogeologic
Data Analysis, John Wiley, New York, 1998. Example 13-2, Pages 689-692
}
\usage{
neuzil
}
\description{
The information for this slug test is:
\itemize{
\item Radius of the well (in m): rw <- 0.067
\item Effective compressibility (in Pa-1): Ceff <- 2.723e-09
\item Volume of the test section (in m3): Vs <- 1.59
}
}
\keyword{datasets}
