% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuzil_utilities.R
\name{neuzil_calculate_parameters}
\alias{neuzil_calculate_parameters}
\title{neuzil_calculate_parameters}
\usage{
neuzil_calculate_parameters(ptest, par, hydraulic = TRUE)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{par}{A list with the wellbore storage coefficient (cd) and t0 parameters}

\item{hydraulic}{Logical flag to indicate if hydraulic parameters are calculated.
If False, the the statistcal parameter (a and t0) are calculated.}
}
\value{
A list with
\itemize{
\item Tr: Transmissivity
\item Ss: Storage coefficient
\item radius_influence
}
}
\description{
Function to calculate the hydraulic parameters for the Neuzil solution
}
\examples{
data(neuzil)
ptest.neuzil <- pumping_test("Well1", Q = 0, r = 0.0, t = neuzil$t, s = neuzil$s)
par <- list(rw = 0.067, ceff = 2.723e-9, vs = 1.59)
ptest.neuzil$additional_parameters <- par
sol0 <- neuzil_solution_initial(ptest.neuzil)
neuzil.par <- neuzil_calculate_parameters(ptest.neuzil, sol0)
print(neuzil.par)
}
\references{
Neuzil, C. E. On conducting the modified 'Slug' test in tight formations. Water
Resources Research, 1982, 18, 2, 439-441
}
\seealso{
Other neuzil functions: \code{\link{neuzil_WF_LT}},
  \code{\link{neuzil_solution_dlogt}},
  \code{\link{neuzil_solution_initial}},
  \code{\link{neuzil_solution}},
  \code{\link{neuzil_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{neuzil functions}
