% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuzil_utilities.R
\name{neuzil_solution}
\alias{neuzil_solution}
\title{neuzil_solution}
\usage{
neuzil_solution(ptest, cd, t0, t)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{cd}{Value of the wellbore storage coefficient}

\item{t0}{Intercept}

\item{t}{Numeric vector with time}
}
\value{
A numeric vector with the calculated drawdown
}
\description{
Calculates the drawdown using Neuzil solution
}
\examples{
data(neuzil)
ptest.neuzil <- pumping_test("Well1", Q = 0, r = 0.0, t = neuzil$t, s = neuzil$s)
par <- list(rw = 0.067, ceff = 2.723e-9, vs = 1.59)
ptest.neuzil$additional_parameters <- par
sol0 <- neuzil_solution_initial(ptest.neuzil)
sol <- neuzil_solution(ptest.neuzil, sol0$cd, sol0$t0, ptest.neuzil$t)
print(sol)
}
\references{
Neuzil, C. E. On conducting the modified 'Slug' test in tight formations. Water
Resources Research, 1982, 18, 2, 439-441
}
\seealso{
Other neuzil functions: \code{\link{neuzil_WF_LT}},
  \code{\link{neuzil_calculate_parameters}},
  \code{\link{neuzil_solution_dlogt}},
  \code{\link{neuzil_solution_initial}},
  \code{\link{neuzil_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{neuzil functions}
