% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuzil_utilities.R
\name{neuzil_solution_initial}
\alias{neuzil_solution_initial}
\title{neuzil_solution_initial}
\usage{
neuzil_solution_initial(ptest)
}
\arguments{
\item{ptest}{A pumping_test object}
}
\value{
A list with
\itemize{
\item cd: Wellbore storage coefficient
\item t0: Intercept
}
}
\description{
Calculates the initial values of parameters of Neuzil solution
}
\examples{
data(neuzil)
ptest.neuzil <- pumping_test("Well1", Q = 0, r = 0.0, t = neuzil$t, s = neuzil$s)
par <- list(rw = 0.067, ceff = 2.723e-9, vs = 1.59)
ptest.neuzil$additional_parameters <- par
sol0 <- neuzil_solution_initial(ptest.neuzil)
print(sol0)
}
\references{
Neuzil, C. E. On conducting the modified 'Slug' test in tight formations. Water
Resources Research, 1982, 18, 2, 439-441
}
\seealso{
Other neuzil functions: \code{\link{neuzil_WF_LT}},
  \code{\link{neuzil_calculate_parameters}},
  \code{\link{neuzil_solution_dlogt}},
  \code{\link{neuzil_solution}},
  \code{\link{neuzil_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{neuzil functions}
