% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuzil_utilities.R
\name{neuzil_well_function}
\alias{neuzil_well_function}
\title{neuzil_well_function}
\usage{
neuzil_well_function(td, par)
}
\arguments{
\item{td}{Numeric vector with the dimensionless time}

\item{par}{A list with the wellbore storage coefficient (cd) and stehfest coeffs (coeffs)}
}
\value{
A numeric vector with the dimensionless drawdown calculated using the Neuzil solution
}
\description{
Calculates the well function of the Neuzil solution
}
\examples{
td <- logseq(-1, 3, 50)
par <- list(cd = 0.01, coeffs = stehfest_coeffs(8) )
W <- neuzil_well_function(td, par)
plot(td, W, type = "l", log = "xy", main = "NEUZIL SOLUTION: Well function")
}
\references{
Neuzil, C. E. On conducting the modified 'Slug' test in tight formations. Water
Resources Research, 1982, 18, 2, 439-441
}
\seealso{
Other neuzil functions: \code{\link{neuzil_WF_LT}},
  \code{\link{neuzil_calculate_parameters}},
  \code{\link{neuzil_solution_dlogt}},
  \code{\link{neuzil_solution_initial}},
  \code{\link{neuzil_solution}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{neuzil functions}
