% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuzil_utilities.R
\name{neuzil_WF_LT}
\alias{neuzil_WF_LT}
\title{neuzil_WF_LT}
\usage{
neuzil_WF_LT(p, arg1, ...)
}
\arguments{
\item{p}{Laplace parameter}

\item{arg1}{Value of the wellbore storage coefficient (cd)}

\item{...}{Additional parameters}
}
\value{
A numeric vector with the calculated drawdown in the Laplace domain
}
\description{
Calculates the drawdown using Neuzil function in the Laplace domain
}
\examples{
p <- 0.5
cd <- 0.05
s <- neuzil_WF_LT(p, cd)
print(s)
}
\references{
Neuzil, C. E. On conducting the modified 'Slug' test in tight formations. Water
Resources Research, 1982, 18, 2, 439-441
}
\seealso{
Other neuzil functions: \code{\link{neuzil_calculate_parameters}},
  \code{\link{neuzil_solution_dlogt}},
  \code{\link{neuzil_solution_initial}},
  \code{\link{neuzil_solution}},
  \code{\link{neuzil_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{neuzil functions}
