% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest_data.R
\docType{data}
\name{papadopulos_cooper}
\alias{papadopulos_cooper}
\title{estimation of a pumping test in an aquifer with well-bore storage}
\format{A data frame with 46 rows and 2 variables:
\describe{
  \item{t}{time, in seconds}
  \item{s}{drawdown, in meters}
}}
\source{
Hall, P. Water well and aquifer test analysis 1996, Water Resources
Publications, LLC, 412 pp. Pages 171-174.
}
\usage{
papadopulos_cooper
}
\description{
The information for this pumping test is:
\itemize{
 \item Radius of the well (in m): rw <- 0.6096
 \item Radius of the casing (in m): rc <- 0.6096
 \item Distance to pumping well (in m): r <- 3.048
 \item Pumping Rate (in m3/s):  Q <- 0.0050472
}
}
\keyword{datasets}
