% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papadopulos_cooper_utilities.R
\name{papadopulos_cooper_calculate_parameters}
\alias{papadopulos_cooper_calculate_parameters}
\title{papadopulos_cooper_calculate_parameters}
\usage{
papadopulos_cooper_calculate_parameters(ptest, par, hydraulic = TRUE)
}
\arguments{
\item{ptest}{A pumping test object}

\item{par}{A list with the values of a (slope) and t0 (Intercept) of the straight line fitted
to the drawdown data using the Cooper-Jacob approach.}

\item{hydraulic}{Logical flag to indicate if hydraulic parameters are calculated. If False,
the the statistcal parameters (a and t0) are calculated.}
}
\value{
A list with the hydraulic parameters:
\itemize{
\item Tr: Transmissivity
\item Ss: Storage coefficient
\item cd: Wellbore Storage
\item rho: Dimensionless radius
}
}
\description{
Function to calculate the hydraulic parameters of the Papadopulos-Cooper solution
}
\examples{
data(papadopulos_cooper)
ptest <- pumping_test("Well1", Q=0.0050472, r = 3.048, t = as.numeric(papadopulos_cooper$t),
                      s = papadopulos_cooper$s)
par <- list(rw = 0.6096, rc = 0.6096 )
ptest$additional_parameters <- par
sol0 <- papadopulos_cooper_solution_initial(ptest)
pc.par <- papadopulos_cooper_calculate_parameters(ptest, sol0)
print(pc.par)
}
\references{
Papadopulos, I. S. & Cooper, H. H. Drawdown in a well of large diameter. Water
Resources Research, 1967, 3, 241-244.
}
\seealso{
Other papadopulos_cooper functions: \code{\link{papadopulos_cooper_F_WF_LT_dlogt}},
  \code{\link{papadopulos_cooper_F_WF_LT}},
  \code{\link{papadopulos_cooper_solution_dlogt}},
  \code{\link{papadopulos_cooper_solution_initial}},
  \code{\link{papadopulos_cooper_solution}},
  \code{\link{papadopulos_cooper_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{papadopulos_cooper functions}
