% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papadopulos_cooper_utilities.R
\name{papadopulos_cooper_F_WF_LT}
\alias{papadopulos_cooper_F_WF_LT}
\title{papadopulos_cooper_F_WF_LT}
\usage{
papadopulos_cooper_F_WF_LT(p, arg1, arg2)
}
\arguments{
\item{p}{Laplace parameter}

\item{arg1}{Dimensionless wellbore storage coefficient (cd)}

\item{arg2}{Dimensionless radius (rho)}
}
\value{
A numeric vector with the drawdown in the Laplace domain
}
\description{
Function to calculate the drawdown of the Papadopulos-Cooper solution in the Laplace domain
}
\examples{
p <- 0.5
cd <- 1e-2
rho <- 0.5
res <- papadopulos_cooper_F_WF_LT(p, cd, rho)
print(res)
}
\references{
Papadopulos, I. S. & Cooper, H. H. Drawdown in a well of large diameter. Water
Resources Research, 1967, 3, 241-244.
}
\seealso{
Other papadopulos_cooper functions: \code{\link{papadopulos_cooper_F_WF_LT_dlogt}},
  \code{\link{papadopulos_cooper_calculate_parameters}},
  \code{\link{papadopulos_cooper_solution_dlogt}},
  \code{\link{papadopulos_cooper_solution_initial}},
  \code{\link{papadopulos_cooper_solution}},
  \code{\link{papadopulos_cooper_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{papadopulos_cooper functions}
