% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papadopulos_cooper_utilities.R
\name{papadopulos_cooper_F_WF_LT_dlogt}
\alias{papadopulos_cooper_F_WF_LT_dlogt}
\title{papadopulos_cooper_F_WF_LT_dlogt}
\usage{
papadopulos_cooper_F_WF_LT_dlogt(p, arg1, arg2)
}
\arguments{
\item{p}{Laplace parameter}

\item{arg1}{Value of dimensionless wellbore storage coefficient (cd)}

\item{arg2}{Value of dimensionless radius (rho)}
}
\value{
A numeric vector with the derivative of the drawdown with respect to the logarithm
of time for thepapadopulos_cooper solution in the Laplace domain
}
\description{
Calculates the derivative of the drawdown with respect to the logarithm of time for the
papadopulos_cooper solution in the Laplace domain
}
\examples{
data(papadopulos_cooper)
}
\references{
Papadopulos, I. S. & Cooper, H. H. Drawdown in a well of large diameter. Water
Resources Research, 1967, 3, 241-244.
}
\seealso{
Other papadopulos_cooper functions: \code{\link{papadopulos_cooper_F_WF_LT}},
  \code{\link{papadopulos_cooper_calculate_parameters}},
  \code{\link{papadopulos_cooper_solution_dlogt}},
  \code{\link{papadopulos_cooper_solution_initial}},
  \code{\link{papadopulos_cooper_solution}},
  \code{\link{papadopulos_cooper_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{papadopulos_cooper functions}
