% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papadopulos_cooper_utilities.R
\name{papadopulos_cooper_solution_initial}
\alias{papadopulos_cooper_solution_initial}
\title{papadopulos_cooper_solution_initial}
\usage{
papadopulos_cooper_solution_initial(ptest)
}
\arguments{
\item{ptest}{A pumping_test object}
}
\value{
A list with
\itemize{
\item a: Slope of the straight line fitted to the drawdown data using the
Cooper-Jacob approach
\item t0: Intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach
}
}
\description{
Function to calculate the initial solution of the Papadopulos-Cooper model
}
\examples{
data(papadopulos_cooper)
ptest <- pumping_test("Well1", Q=0.0050472, r = 3.048, t = as.numeric(papadopulos_cooper$t),
                      s = papadopulos_cooper$s)
sol0 <- papadopulos_cooper_solution_initial(ptest)
print(sol0)
}
\references{
Papadopulos, I. S. & Cooper, H. H. Drawdown in a well of large diameter. Water
Resources Research, 1967, 3, 241-244.
}
\seealso{
Other papadopulos_cooper functions: \code{\link{papadopulos_cooper_F_WF_LT_dlogt}},
  \code{\link{papadopulos_cooper_F_WF_LT}},
  \code{\link{papadopulos_cooper_calculate_parameters}},
  \code{\link{papadopulos_cooper_solution_dlogt}},
  \code{\link{papadopulos_cooper_solution}},
  \code{\link{papadopulos_cooper_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{papadopulos_cooper functions}
