% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papadopulos_cooper_utilities.R
\name{papadopulos_cooper_well_function}
\alias{papadopulos_cooper_well_function}
\title{papadopulos_cooper_well_function}
\usage{
papadopulos_cooper_well_function(td, par)
}
\arguments{
\item{td}{Dimensionless time}

\item{par}{List with the  \eqn{cd} and \eqn{\rho} parameters.
The well storage (\eqn{cd} parameter) is defined as
\deqn{cd = \frac{r_{w}^{2} S}{r_{c}^{2}}}
where \eqn{r_{w}} is the radius of the well, \eqn{r_{c}} is the radius of the casing,
and \eqn{S} is the storage coefficient. The dimensionless radius (\eqn{\rho} parameter)
 is defined as:
\deqn{\rho=\frac{r}{r_{w}}}
where \eqn{r} is the distance to the observation well, and \eqn{r_{w}} is the radius
of the well.}
}
\value{
The value of the well function for the Papadopulos-Cooper model
}
\description{
Function to calculate the well function of the Papadopulos-Cooper solution that describe the drawdown
variations for large diameter wells
}
\examples{
td <- logseq(-2,4,30)
par <- list(cd = 0.1, rho = 0.1)
res <- papadopulos_cooper_well_function(td, par)
}
\references{
Papadopulos, I. S. & Cooper, H. H. Drawdown in a well of large diameter. Water
Resources Research, 1967, 3, 241-244.
}
\seealso{
Other papadopulos_cooper functions: \code{\link{papadopulos_cooper_F_WF_LT_dlogt}},
  \code{\link{papadopulos_cooper_F_WF_LT}},
  \code{\link{papadopulos_cooper_calculate_parameters}},
  \code{\link{papadopulos_cooper_solution_dlogt}},
  \code{\link{papadopulos_cooper_solution_initial}},
  \code{\link{papadopulos_cooper_solution}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{papadopulos_cooper functions}
