% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest.R
\name{plot_model_diagnostic}
\alias{plot_model_diagnostic}
\title{plot_model_diagnostic}
\usage{
plot_model_diagnostic(ptest, ...)
}
\arguments{
\item{ptest}{A pumping_test object. It must be estimated.}

\item{...}{Additional parameters to the plot function used in scatter.smooth}
}
\description{
Function to plot the residuals of an estimated pumping test
}
\examples{
data(theis)
ptest <- pumping_test("Test", Q = 1.388e-2, r = 250, t = theis$t, s = theis$s)
ptest.fit <- fit(ptest, "theis")
hydraulic.parameters(ptest) <- ptest.fit$hydraulic_parameters
fit.parameters(ptest) <- ptest.fit$parameters
model(ptest) <- "theis"
estimated(ptest) <- TRUE
plot(ptest, type = 'estimation', dmethod = "spline", d = 30)
plot_model_diagnostic(ptest)
}
\seealso{
Other base functions: \code{\link{additional.parameters<-}},
  \code{\link{confint.pumping_test}},
  \code{\link{confint_bootstrap}},
  \code{\link{confint_jackniffe}},
  \code{\link{confint_wald}}, \code{\link{estimated<-}},
  \code{\link{evaluate}}, \code{\link{fit.optimization}},
  \code{\link{fit.parameters<-}},
  \code{\link{fit.sampling}}, \code{\link{fit}},
  \code{\link{hydraulic.parameter.names<-}},
  \code{\link{hydraulic.parameters<-}},
  \code{\link{model.parameters}}, \code{\link{model<-}},
  \code{\link{plot.pumping_test}},
  \code{\link{print.pumping_test}},
  \code{\link{pumping_test}}, \code{\link{simulate}},
  \code{\link{summary.pumping_test}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{base functions}
