% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest.R
\name{plot_sample_influence}
\alias{plot_sample_influence}
\title{plot_sample_influence}
\usage{
plot_sample_influence(res, ...)
}
\arguments{
\item{res}{A list with the results of the jackniffe CI estimation.}

\item{...}{Additional parameters to the plot function}
}
\value{
A plot with the influence measure of each sample
}
\description{
Function to create a plot with an influence measure for each sample of the pumping test.
This plot is helpful in the identification of the samples that are influential in the
estimation of the hydraulic parameters.
}
\seealso{
Other base functions: \code{\link{additional.parameters<-}},
  \code{\link{confint.pumping_test}},
  \code{\link{confint_bootstrap}},
  \code{\link{confint_jackniffe}},
  \code{\link{confint_wald}}, \code{\link{estimated<-}},
  \code{\link{evaluate}}, \code{\link{fit.optimization}},
  \code{\link{fit.parameters<-}},
  \code{\link{fit.sampling}}, \code{\link{fit}},
  \code{\link{hydraulic.parameter.names<-}},
  \code{\link{hydraulic.parameters<-}},
  \code{\link{model.parameters}}, \code{\link{model<-}},
  \code{\link{plot.pumping_test}},
  \code{\link{plot_model_diagnostic}},
  \code{\link{plot_uncert}},
  \code{\link{print.pumping_test}},
  \code{\link{pumping_test}}, \code{\link{simulate}},
  \code{\link{summary.pumping_test}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{base functions}
