% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest.R
\name{plot_uncert}
\alias{plot_uncert}
\title{plot_uncert}
\usage{
plot_uncert(ptest, cex = 1, ...)
}
\arguments{
\item{ptest}{A pumping_test object.}

\item{cex}{Character expansion}

\item{...}{Additional parameters for the plot function}
}
\description{
Function to plot the distributions of the confidence intervals estimated using bootstrapp
from a model fitted using nonlinear regression.
}
\seealso{
Other base functions: \code{\link{additional.parameters<-}},
  \code{\link{confint.pumping_test}},
  \code{\link{confint_bootstrap}},
  \code{\link{confint_jackniffe}},
  \code{\link{confint_wald}}, \code{\link{estimated<-}},
  \code{\link{evaluate}}, \code{\link{fit.optimization}},
  \code{\link{fit.parameters<-}},
  \code{\link{fit.sampling}}, \code{\link{fit}},
  \code{\link{hydraulic.parameter.names<-}},
  \code{\link{hydraulic.parameters<-}},
  \code{\link{model.parameters}}, \code{\link{model<-}},
  \code{\link{plot.pumping_test}},
  \code{\link{plot_model_diagnostic}},
  \code{\link{plot_sample_influence}},
  \code{\link{print.pumping_test}},
  \code{\link{pumping_test}}, \code{\link{simulate}},
  \code{\link{summary.pumping_test}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{base functions}
