% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquifer_utilities.R
\name{print.aquifer}
\alias{print.aquifer}
\title{print.aquifer}
\usage{
\method{print}{aquifer}(x, ...)
}
\arguments{
\item{x}{An aquifer object}

\item{...}{Additional parameters}
}
\value{
This function prints all the elements of the aquifer on the console
}
\description{
Function to print the aquifer object on the console
}
\seealso{
Other aquifer functions: \code{\link{add.well<-}},
  \code{\link{aquifer}}, \code{\link{calculate_drawdown}},
  \code{\link{summary.aquifer}}
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{aquifer functions}
