% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest.R
\name{print.pumping_test}
\alias{print.pumping_test}
\title{print}
\usage{
\\method{print}{pumping_test}(x, ...)
}
\arguments{
\item{x}{A pumping_test object}

\item{...}{Additional parameters to the data.frame print function}
}
\value{
This function prints the information from the pumping test on the screen
}
\description{
Function to print on the screen a pumping_test object
}
\examples{
# Define a pumping test
tt <- logseq(0, 5, 30)
Tr <- 1e-3
Ss <- 0.0001
r <- 1e3
Q <- 1e-3
s <- Q/(4*pi*Tr)*theis_well_function(tt)
ptest <- pumping_test("Test", Q = Q, r = r, t = tt, s = s)
print(ptest)
}
\seealso{
Other base functions: \code{\link{additional.parameters<-}},
  \code{\link{confint.pumping_test}},
  \code{\link{confint_bootstrap}},
  \code{\link{confint_jackniffe}},
  \code{\link{confint_wald}}, \code{\link{estimated<-}},
  \code{\link{evaluate}}, \code{\link{fit.optimization}},
  \code{\link{fit.parameters<-}},
  \code{\link{fit.sampling}}, \code{\link{fit}},
  \code{\link{hydraulic.parameter.names<-}},
  \code{\link{hydraulic.parameters<-}},
  \code{\link{model.parameters}}, \code{\link{model<-}},
  \code{\link{plot.pumping_test}},
  \code{\link{plot_model_diagnostic}},
  \code{\link{plot_sample_influence}},
  \code{\link{plot_uncert}}, \code{\link{pumping_test}},
  \code{\link{simulate}},
  \code{\link{summary.pumping_test}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{base functions}
