% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptative_mcmc_utilities.R
\name{prior}
\alias{prior}
\title{prior}
\usage{
prior(par, prior.pdf, prior.parameters)
}
\arguments{
\item{par}{A numeric vector with the values of the drawdown model parameters}

\item{prior.pdf}{A character vector with the names of the probability density functions for
each parameters. Currently only 'unif' and 'norm' are supported.}

\item{prior.parameters}{A matrix with the parameters of the probability density functions. If
the pdf is 'unif' then the min and max must be specified, whereas in the case of a normal
distribution the mean and standard deviation.}
}
\value{
A numeric value with the sum of the logarithms of the prior distributions
}
\description{
Function to calculate the prior distribution of the parameters used in the drawdown calculations
with analytical models.
}
\seealso{
Other amcmc_auxiliary_function functions: \code{\link{posterior}},
  \code{\link{proposalfunction_cov}},
  \code{\link{proposalfunction}},
  \code{\link{run_adap_metropolis_MCMC}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{amcmc_auxiliary_function functions}
