% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptative_mcmc_utilities.R
\name{proposalfunction}
\alias{proposalfunction}
\title{proposalfunction}
\usage{
proposalfunction(param, proposal.sigma)
}
\arguments{
\item{param}{A numeric vector with the mean value of the parameters.}

\item{proposal.sigma}{A numeric vector with the variance of the parameters.}
}
\value{
A vector with realizations of a multivariate normal distribution with a given mean and
varianced
}
\description{
Function to propose new parameter values from a multivariate normal distribution with a
given mean and variance.
}
\seealso{
Other amcmc_auxiliary_function functions: \code{\link{posterior}},
  \code{\link{prior}}, \code{\link{proposalfunction_cov}},
  \code{\link{run_adap_metropolis_MCMC}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{amcmc_auxiliary_function functions}
