% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptative_mcmc_utilities.R
\name{proposalfunction_cov}
\alias{proposalfunction_cov}
\title{proposalfunction_cov}
\usage{
proposalfunction_cov(mn, cv)
}
\arguments{
\item{mn}{A numeric vector with the mean value of the parameters}

\item{cv}{The covariance matrix}
}
\value{
A vector with realizations of a multivariate normal distribution with a given mean and
covariance
}
\description{
Function to propose new parameter values from a multivariate normal distribution with a
given mean and covariance matrix. This function is used to improve the convergence of a
MCMC.
}
\seealso{
Other amcmc_auxiliary_function functions: \code{\link{posterior}},
  \code{\link{prior}}, \code{\link{proposalfunction}},
  \code{\link{run_adap_metropolis_MCMC}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{amcmc_auxiliary_function functions}
