% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalent_time_utilities.R
\name{recovery_equivalent_time}
\alias{recovery_equivalent_time}
\title{recovery_equivalent_time}
\usage{
recovery_equivalent_time(t_recovery, s_recovery, pump_history)
}
\arguments{
\item{t_recovery}{Vector with the times since recovery started}

\item{s_recovery}{Vector with the drawdowns since recovery started}

\item{pump_history}{Vector with the pump history of the test}
}
\value{
A list with
\itemize{
\item ta equivalent time
\item sa drawdown corresponding to equivalent time
}
}
\description{
Function to calculate the equivalent time of the recovery tests
}
\examples{
data(agarwal_recovery)
pump_history <- 14400
agarwal_corr <- recovery_equivalent_time(agarwal_recovery$t,
                                         agarwal_recovery$s,
                                         pump_history)
}
\references{
Agarwal, R. (1980). A New Method To Account For Producing Time Effects When Drawdown
Type Curves Are Used To Analyze Pressure Buildup And Other Test Data. SPE Paper 9289.
}
\seealso{
Other equivalent_time functions: \code{\link{variable_rate_equivalent_time}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{equivalent_time functions}
