% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_function_utilities.R
\name{residual_sum_squares}
\alias{residual_sum_squares}
\title{residual_sum_squares}
\usage{
residual_sum_squares(par, ptest, model)
}
\arguments{
\item{par}{A list with the parameters of the model.}

\item{ptest}{A pumping_test object.}

\item{model}{A character string with the name of the model}
}
\value{
The numeric value of the residual sum of squares
}
\description{
Function to calculate the residual sum of squares between the measured and simulated drawdown.
}
\seealso{
Other objective_function functions: \code{\link{box_cox_transf}},
  \code{\link{loglikelihood_bc}},
  \code{\link{loglikelihood}},
  \code{\link{max_absolute_deviation}},
  \code{\link{mean_absolute_deviation}},
  \code{\link{mean_square_error}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{objective_function functions}
