% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptative_mcmc_utilities.R
\name{run_adap_metropolis_MCMC}
\alias{run_adap_metropolis_MCMC}
\title{run_adap_metropolis_MCMC}
\usage{
run_adap_metropolis_MCMC(startvalue, iterations = 10000,
  iter_update_par = 100, ptest, model, prior.pdf, prior.parameters,
  proposal.sigma, cov.corr = FALSE)
}
\arguments{
\item{startvalue}{A numeric vector with the fit parameters of the pumping test.}

\item{iterations}{An integer with the number of iterations to run the chain.}

\item{iter_update_par}{An integet specifying the number of iterations to update the
covariance matrix.}

\item{ptest}{A pumping_test object.}

\item{model}{A character string with the name of the model used in the parameter estimation.}

\item{prior.pdf}{A character vector with the distributions of the fit parameters (
'unif' and 'norm' are currently supported).}

\item{prior.parameters}{A matrix with the parameters of the distributions (min and max for
uniform distributions, mean and sd for normal distributions)}

\item{proposal.sigma}{A numeric vector with the standard deviations of the proposal distribution.}

\item{cov.corr}{A logical flag indicating if the covariance matrix must be corrected for positive
definiteness.}
}
\value{
A matrix with the sampled values of the fit parameters.
}
\description{
Function to sample for a complex probabibility density function using MCMC with the
adaptative Metropolis algorithm proposed by Roberts and Rosenthal(2009).
}
\details{
This function implements the adaptative MCMC proposed by Roberts and Rosenthal (2009), in
which the proposal distribution \eqn{Q_{n}(x, \cdot)} is given by:
\deqn{Q_{n}(x, \cdot) = \left\{
\begin{aligned}
&(1-\theta)N(x, (2.38)^{2}\Sigma_{n}/d) + \theta N(x,(0.1)^{2}I_{d}/d), &\Sigma_{n}\text{ is positive definite} \\
&N(x,(0.1)^{2}I_{d}), &\Sigma_{n}\text{ is not positive definitive}\\
\end{aligned}
\right.}
where
\itemize{
\item \eqn{\theta \in (0,1)}: control parameters
\item \eqn{N()}: Normal distribution
\item \eqn{\Sigma_{n}}: empirical covariance matrix
\item \eqn{d}: number of parameters
\item \eqn{I_{d}}: identity matrix of size \eqn{d}.
}
This proposal function is implemented in the function proposalfunction_cov.
}
\references{
Roberts, G. O. & Rosenthal, J. S. Examples of adaptive MCMC Journal of Computational
 and Graphical Statistics, 2009, 18, 349-367.
}
\seealso{
Other amcmc_auxiliary_function functions: \code{\link{posterior}},
  \code{\link{prior}}, \code{\link{proposalfunction_cov}},
  \code{\link{proposalfunction}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{amcmc_auxiliary_function functions}
