% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twalk_utilities.R
\name{run_twalk_MCMC}
\alias{run_twalk_MCMC}
\title{run_twalk_MCMC}
\usage{
run_twalk_MCMC(startvalue1, startvalue2, ptest, model, prior.pdf,
  prior.parameters, iterations = 10000)
}
\arguments{
\item{startvalue1}{Numeric vector with the values of the hydraulic parameters used to initialize the chain.}

\item{startvalue2}{Numeric vector with the values of the hydraulic parameters used to initialize the chain.
It must be different than the vector startvalue1.}

\item{ptest}{A pumping test object}

\item{model}{A character string specifying the model that describe the flow during the pumping test}

\item{prior.pdf}{A character vector with the names of the PDF of the hydraulic parameters.
Currently only uniform ('unif') and normal ('norm') distributions are supported.}

\item{prior.parameters}{A numeric vector with the parameters of the PDF of the hydraulic parameters.
The min and max values of the hydraulic parameters are specified in the case of a uniform distribution.
The mean and sigma values of the hydraulic parameters are specified in the case of a normal distribution.}

\item{iterations}{Number of iterations to run the chain.}
}
\value{
This function returns a list with the following entries:
\itemize{
\item xxp = xxp
\item logdensity: matrix with the values of the logdensity of the accepted parameters.
\item acceptance: the acceptance rate
}
}
\description{
Function that implements the transverse walk sampling strategy to obtain realizations of the
posterior distribution of the pumping test parameters.
}
\references{
Christen, J. A. & Fox, C. A general purpose sampling algorithm for continuous distributions (the t-walk).
Bayesian Analysis, 2010, 5, 2, 263-281.
}
\seealso{
Other twalk_auxiliary_function functions: \code{\link{blow_kernel}},
  \code{\link{hop_kernel}},
  \code{\link{log_ratio_density_blow}},
  \code{\link{log_ratio_density_hop}}, \code{\link{psi1}},
  \code{\link{traverse_kernel}}, \code{\link{walk_kernel}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{twalk_auxiliary_function functions}
