% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest.R
\name{simulate}
\alias{simulate}
\title{simulate}
\usage{
simulate(model, Q, r, t, par, add.par = NULL, sigma = 0.01,
  seed = 12345)
}
\arguments{
\item{model}{Model used to generate the drawdown curve}

\item{Q}{Pumping rate in m3/s}

\item{r}{Distance from pumping well to observation well in m}

\item{t}{Numeric vector with the values of time to simulate the drawdown}

\item{par}{List with parameters of the model (Transmissivity, Storage Coefficient, etc)}

\item{add.par}{List with the additional parameter required to define the pumping_test object}

\item{sigma}{Standard deviation of the noise added to the simulated drawdown.}

\item{seed}{Seed for the random number generator}
}
\value{
A pumping_test object with the simulated drawdown
}
\description{
Function to generate a drawdown curve using a given model with given hydraulic parameters
and noise
}
\examples{
 par <- list(Tr = 1.4e-3, Ss = 2.1e-5)
 t <- logseq(2, 5, 50)
 ptest.sim <- simulate('theis', Q = 1.388e-2, r =250, t = t, par)
}
\seealso{
Other base functions: \code{\link{additional.parameters<-}},
  \code{\link{confint.pumping_test}},
  \code{\link{confint_bootstrap}},
  \code{\link{confint_jackniffe}},
  \code{\link{confint_wald}}, \code{\link{estimated<-}},
  \code{\link{evaluate}}, \code{\link{fit.optimization}},
  \code{\link{fit.parameters<-}},
  \code{\link{fit.sampling}}, \code{\link{fit}},
  \code{\link{hydraulic.parameter.names<-}},
  \code{\link{hydraulic.parameters<-}},
  \code{\link{model.parameters}}, \code{\link{model<-}},
  \code{\link{plot.pumping_test}},
  \code{\link{plot_model_diagnostic}},
  \code{\link{print.pumping_test}},
  \code{\link{pumping_test}},
  \code{\link{summary.pumping_test}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{base functions}
