% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_utilities.R
\name{smoothing.pumping_test}
\alias{smoothing.pumping_test}
\title{smoothing.pumping_test}
\usage{
smoothing.pumping_test(ptest, ntvalues = NULL, method = c("spline",
  "kernsmooth", "lokern", "locpol", "lpridge"))
}
\arguments{
\item{ptest}{A pumping_test object}

\item{ntvalues}{Integer, number of values}

\item{method}{Character string specifying the method to be used in the smoothing.
Current options include:
\itemize{
\item spline: smooth.spline
\item kernsmooth: KernSmooth
\item lokern: local kernel regression
\item locpoly: local polynomial regression
\item lpridge: lpridge
\item wavelet: Wavelet thresholding
}}
}
\description{
Function to smooth or filter the drawdown data using different methodologies.
}
