% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace_inversion_utilities.R
\name{stehfest_coeffs}
\alias{stehfest_coeffs}
\title{stehfest_coeffs}
\usage{
stehfest_coeffs(n)
}
\arguments{
\item{n}{Number of coefficients}
}
\value{
Numeric vector with the coefficients
}
\description{
Function to calculate the coefficients used in the stehfest inversion. These coefficients
\eqn{c_{i}}are given by:
\deqn{c_{i}=(-1)^{i+n/2} \sum_{k=\left[ \frac{i+1}{2} \right]}^{\min{(i.n/2)}} \frac{k^{n/2}(2k)!}{\left(\frac{n}{2}-k\right)!k!(k-1)!(i-k)!(2k-i)!}}
}
\examples{
coeffs8 <- stehfest_coeffs(8)
coeffs20 <- stehfest_coeffs(20)
}
\seealso{
Other laplace functions: \code{\link{stehfest_inversion}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{laplace functions}
