% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace_inversion_utilities.R
\name{stehfest_inversion}
\alias{stehfest_inversion}
\title{stehfest_inversion}
\usage{
stehfest_inversion(t, coeffs, fun, arg1 = NULL, arg2 = NULL,
  arg3 = NULL)
}
\arguments{
\item{t}{Numeric vector with the time}

\item{coeffs}{Numeric vector with the coefficients}

\item{fun}{function to be used}

\item{arg1, arg2, arg3}{Optional arguments to be used by fun}
}
\value{
Numeric vector
}
\description{
Function to calculate the Laplace inverse transform using the
stehfest algorithm
}
\examples{
test <- function(p,...){1/p}
coeffs8 <- stehfest_coeffs(8)
td <- logseq(-2,4,30)
stehfest_inversion(td, coeffs8,test)
}
\seealso{
Other laplace functions: \code{\link{stehfest_coeffs}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{laplace functions}
