% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquifer_utilities.R
\name{summary.aquifer}
\alias{summary.aquifer}
\title{summary.aquifer}
\usage{
\method{summary}{aquifer}(object, ...)
}
\arguments{
\item{object}{An aquifer object}

\item{...}{Additional arguments}
}
\value{
This function prints a summary of the aquifer object on the screen
}
\description{
Function to print a summary of an aquifer object
}
\seealso{
Other aquifer functions: \code{\link{add.well<-}},
  \code{\link{aquifer}}, \code{\link{calculate_drawdown}},
  \code{\link{print.aquifer}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{aquifer functions}
