% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest.R
\name{summary.pumping_test}
\alias{summary.pumping_test}
\title{summary.pumping_test}
\usage{
\method{summary}{pumping_test}(object, ...)
}
\arguments{
\item{object}{A pumping_test object}

\item{...}{additional parameters for the data.frame summary function}
}
\value{
This function displays on the console a summary of the drawdown data
}
\description{
Function to display a short summary of the drawdown data
}
\examples{
# Define a pumping test
tt <- logseq(0, 5, 30)
Tr <- 1e-3
Ss <- 0.0001
r <- 1e3
Q <- 1e-3
s <- Q/(4*pi*Tr)*theis_well_function(tt)
ptest <- pumping_test("Test", Q = Q, r = r, t = tt, s = s)
summary(ptest)
}
\seealso{
Other base functions: \code{\link{additional.parameters<-}},
  \code{\link{confint.pumping_test}},
  \code{\link{confint_bootstrap}},
  \code{\link{confint_jackniffe}},
  \code{\link{confint_wald}}, \code{\link{estimated<-}},
  \code{\link{evaluate}}, \code{\link{fit.optimization}},
  \code{\link{fit.parameters<-}},
  \code{\link{fit.sampling}}, \code{\link{fit}},
  \code{\link{hydraulic.parameter.names<-}},
  \code{\link{hydraulic.parameters<-}},
  \code{\link{model.parameters}}, \code{\link{model<-}},
  \code{\link{plot.pumping_test}},
  \code{\link{plot_model_diagnostic}},
  \code{\link{print.pumping_test}},
  \code{\link{pumping_test}}, \code{\link{simulate}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{base functions}
