% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest_data.R
\docType{data}
\name{td05}
\alias{td05}
\title{Cooper Slug Test - Initial solution}
\format{A data frame with 100 rows and 1 variable:
\describe{
  \item{td05}{td05}
}}
\usage{
td05
}
\description{
Cooper Slug Test - Initial solution
}
\keyword{datasets}
