% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest_data.R
\docType{data}
\name{theis}
\alias{theis}
\title{Pumping test data from a confined aquifer.}
\format{A data frame with 22 rows and 2 variables:
\describe{
  \item{t}{time, in seconds}
  \item{s}{drawdown, in meters}
}}
\source{
C.W. Fetter, 2001, Applied Hydrogeology, Fourth Edition. Prentice Hall,
Upper Saddle River, 598 pp. Table 5.1, page 172
}
\usage{
theis
}
\description{
A dataset containing the time and drawdowns of a pumping test in a confined aquifer.
The information for this pumping test is:
\itemize{
\item Pumping rate (in m3/s): Q <- 1.3888e-2
\item Distance to pumping well (in m): r <- 250
}
}
\keyword{datasets}
