% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theis_utilities.R
\name{theis_calculate_parameters}
\alias{theis_calculate_parameters}
\title{theis_calculate_parameters}
\usage{
theis_calculate_parameters(ptest, par, hydraulic = TRUE)
}
\arguments{
\item{ptest}{An object of the pumping_test class}

\item{par}{List with the values of a and t0 obtained using the theis_solution_initial}

\item{hydraulic}{Logical flag to indicate if hydraulic parameters are calculated. If False, the the statistcal parameter are calculated.
function}
}
\value{
A list with
\itemize{
  \item Tr: Transmisivity
  \item Ss: Storage coefficient
  \item radius_influence
}
}
\description{
Function to calculate the Transmisivity, Storage Coefficient and Radius of influence
for the Theis model
}
\examples{
# Data from a confined aquifer
data(theis)
ptest <- pumping_test('Well1', Q = 1.3e-3, r = 200, t = theis$t, s = theis$s)
res <- theis_solution_initial(ptest)
# Calculate hydraulic parameters from fit parameters
theis_hydrpar <- theis_calculate_parameters(ptest, res)
print(theis_hydrpar)
# Calculate fit parameters from hydraulic parameters
res <- list(Tr = 1.5e-3, Ss = 2e-5)
theis_fitpar <- theis_calculate_parameters(ptest, res, FALSE)
}
\references{
Theis, C. The relation between the lowering of the piezometric surface and the rate
and duration of discharge of a well using groundwater storage. Transactions of the
American Geophysical Union, 1935, 2, 519-524
}
\seealso{
Other theis functions: \code{\link{exponential_integral}},
  \code{\link{theis_WF_LT_dlogt}},
  \code{\link{theis_WF_LT}},
  \code{\link{theis_solution_dlogt}},
  \code{\link{theis_solution_initial}},
  \code{\link{theis_solution}},
  \code{\link{theis_well_function_large}},
  \code{\link{theis_well_function_small}},
  \code{\link{theis_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{theis functions}
