% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theis_utilities.R
\name{theis_solution}
\alias{theis_solution}
\title{theis_solution}
\usage{
theis_solution(ptest, a, t0, t)
}
\arguments{
\item{ptest}{A pumpint test object}

\item{a}{slope of the straight line fitted to the drawdown data using the
Cooper-Jacob approach}

\item{t0}{intercept of the straight line fitted to the drawdown data using the
Cooper-Jacob approach}

\item{t}{Numeric vector with the time values}
}
\value{
A numeric vector with the drawdown
}
\description{
Function to calculate the drawdown using the Theis solution
}
\examples{
# Data from a confined aquifer
data(theis)
ptest <- pumping_test('Well1', Q = 1.3e-3, r = 200, t = theis$t, s = theis$s)
res <- theis_solution_initial(ptest)
s_model <- theis_solution(ptest,res$a, res$t0, theis$t)
#Compare drawdown from model with measurements
mse <- sqrt(mean((s_model-theis$s)^2))
print(mse)
}
\references{
Theis, C. The relation between the lowering of the piezometric surface and the rate
and duration of discharge of a well using groundwater storage. Transactions of the
American Geophysical Union, 1935, 2, 519-524
}
\seealso{
Other theis functions: \code{\link{exponential_integral}},
  \code{\link{theis_WF_LT_dlogt}},
  \code{\link{theis_WF_LT}},
  \code{\link{theis_calculate_parameters}},
  \code{\link{theis_solution_dlogt}},
  \code{\link{theis_solution_initial}},
  \code{\link{theis_well_function_large}},
  \code{\link{theis_well_function_small}},
  \code{\link{theis_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{theis functions}
