% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theis_utilities.R
\name{theis_well_function}
\alias{theis_well_function}
\title{theis_well_function}
\usage{
theis_well_function(td, par = NULL)
}
\arguments{
\item{td}{Dimensionless time \eqn{t_n} defined as
\deqn{ t_d=\frac{4Tt}{r^{2}S} }
where \eqn{r} is the distance between the pumping and observation well in m, \eqn{S} is the
storage coefficient (dimensionless), \eqn{T} is the Transmisivity in m2/s and \eqn{t} is the time in s}

\item{par}{additional parameters}
}
\value{
This function returns the value of the Theis well function
}
\description{
Function to evaluate the Theis well function:
\deqn{ W(u) = \int_{u}^{\infty} \frac{\exp{-u}}{u} du}
}
\examples{
u <- seq(0.05,30,0.05)
td <- 1.0/u
W <- theis_well_function(td, par)
plot(td, W, type = "l", log = "xy")
}
\references{
Theis, C. The relation between the lowering of the piezometric surface and the rate
and duration of discharge of a well using groundwater storage. Transactions of the
American Geophysical Union, 1935, 2, 519-524
}
\seealso{
Other theis functions: \code{\link{exponential_integral}},
  \code{\link{theis_WF_LT_dlogt}},
  \code{\link{theis_WF_LT}},
  \code{\link{theis_calculate_parameters}},
  \code{\link{theis_solution_dlogt}},
  \code{\link{theis_solution_initial}},
  \code{\link{theis_solution}},
  \code{\link{theis_well_function_large}},
  \code{\link{theis_well_function_small}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{theis functions}
