% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theis_utilities.R
\name{theis_well_function_small}
\alias{theis_well_function_small}
\title{theis_well_function_small}
\usage{
theis_well_function_small(u)
}
\arguments{
\item{u}{Dimensionless \eqn{u} variable defined as:
\deqn{ u=\frac{r^{2}S}{4Tt} }
where \eqn{r} is the distance between the pumping and observation well in m, \eqn{S} is the
storage coefficient (dimensionless), \eqn{T} is the Transmisivity in m2/s and \eqn{t} is the time in s}
}
\value{
This function returns the value of the dimensionless drawdown
}
\description{
Function to calculate the theis well function for values of \eqn{u < 1}
}
\examples{
u <- seq(0.01, 1, 0.01)
Ws <- theis_well_function_small(u)
plot(1/u, Ws, type = "l", log = "xy")
}
\references{
Theis, C. The relation between the lowering of the piezometric surface and the rate
and duration of discharge of a well using groundwater storage. Transactions of the
American Geophysical Union, 1935, 2, 519-524
}
\seealso{
Other theis functions: \code{\link{exponential_integral}},
  \code{\link{theis_WF_LT_dlogt}},
  \code{\link{theis_WF_LT}},
  \code{\link{theis_calculate_parameters}},
  \code{\link{theis_solution_dlogt}},
  \code{\link{theis_solution_initial}},
  \code{\link{theis_solution}},
  \code{\link{theis_well_function_large}},
  \code{\link{theis_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{theis functions}
