% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theis_utilities.R
\name{theis_WF_LT}
\alias{theis_WF_LT}
\title{theis_WF_LT}
\usage{
theis_WF_LT(p, ...)
}
\arguments{
\item{p}{Laplace parameter}

\item{...}{Additional parameters}
}
\value{
A numeric vector with the calculated drawdown in Laplace domain
}
\description{
Function to calculate the drawdown in the Laplace domain
}
\examples{
coeffs <- stehfest_coeffs(8)
td <- logseq(-2,4,50)
W <- stehfest_inversion(td, coeffs, theis_WF_LT)
plot(td, W, log = "xy")
}
\references{
Theis, C. The relation between the lowering of the piezometric surface and the rate
and duration of discharge of a well using groundwater storage. Transactions of the
American Geophysical Union, 1935, 2, 519-524
}
\seealso{
Other theis functions: \code{\link{exponential_integral}},
  \code{\link{theis_WF_LT_dlogt}},
  \code{\link{theis_calculate_parameters}},
  \code{\link{theis_solution_dlogt}},
  \code{\link{theis_solution_initial}},
  \code{\link{theis_solution}},
  \code{\link{theis_well_function_large}},
  \code{\link{theis_well_function_small}},
  \code{\link{theis_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail}
}
\concept{theis functions}
