% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twalk_utilities.R
\name{traverse_kernel}
\alias{traverse_kernel}
\title{traverse_kernel}
\usage{
traverse_kernel(x, xp, phi)
}
\arguments{
\item{x}{Value}

\item{xp}{Value}

\item{phi}{Logical vector with the parameters to change.}
}
\value{
A numeric value
}
\description{
PDF of the traverse move
}
\references{
Christen, J. A. & Fox, C. A general purpose sampling algorithm for continuous distributions (the t-walk).
Bayesian Analysis, 2010, 5, 2, 263-281.
}
\seealso{
Other twalk_auxiliary_function functions: \code{\link{blow_kernel}},
  \code{\link{hop_kernel}},
  \code{\link{log_ratio_density_blow}},
  \code{\link{log_ratio_density_hop}}, \code{\link{psi1}},
  \code{\link{run_twalk_MCMC}}, \code{\link{walk_kernel}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{twalk_auxiliary_function functions}
