% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest_data.R
\docType{data}
\name{variable_rate}
\alias{variable_rate}
\title{Data from a pumping test with variable ppumping rates}
\format{A data frame with 18 rows and 2 variables
\describe{
\item{t}{time in seconds}
\item{s}{Drawdown in m}
}}
\source{
Kruseman and de Ridder. 1994. Analysis and Evaluation of pumping test
data. Second Edition pp. 185.
}
\usage{
variable_rate
}
\description{
The information of this pumping test is:
\itemize{
\item Distance to observation well (r): 5 m
\item Pumping rate history:
- Step 1: 300s-1800s, 0.00578 m3/s
- Step 2: 2100s-4800s, 0.0081 m3/s
- Step 3: 5400s-7800s, 0.0069 m3/s
}
}
\keyword{datasets}
