% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalent_time_utilities.R
\name{variable_rate_equivalent_time}
\alias{variable_rate_equivalent_time}
\title{variable_rate_equivalent_time}
\usage{
variable_rate_equivalent_time(t, s, pump_history)
}
\arguments{
\item{t}{Vector with the times since pumping started}

\item{s}{Vector with the drawdowns since pumping started}

\item{pump_history}{Vector with the pump history of the test}
}
\value{
A list with
\itemize{
\item ta equivalent time
\item sa drawdown corresponding to equivalent time
}
}
\description{
Function to calculate the equivalent time required in the variable-rate tests.
}
\references{
Birsoy, Y.K. and W.K. Summers, 1980. Determination of aquifer parameters from step tests
and intermittent pumping, Ground Water, vol. 18, no. 2, pp. 137-146.
}
\seealso{
Other equivalent_time functions: \code{\link{recovery_equivalent_time}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{equivalent_time functions}
