% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpingtest_data.R
\docType{data}
\name{warren_root}
\alias{warren_root}
\title{Pumping test in a double porosity aquifer. estimation with the Warren and Root
model.}
\format{A data frame with 72 rows and 2 variables
\describe{
\item{t}{time in seconds}
\item{s}{Drawdown in m}
}}
\source{
Moench, A. 1984. Double porosity model for a fissured groundwater
reservoir with fracture skin. Water Resources Research, 20(7), 831-846.
Table 2, Page 838.
}
\usage{
warren_root
}
\description{
Fissured tertiary volcanic rocks in the vicinity of the Yucca Mountain
at Nevada Test Site. One pumped well UE-25b#1 and one observation well UE-25a#1.
Pumping duration was about three days. The information of this pumping test is:
\itemize{
\item pumping rate (m3/s): Q <- 3.58e-2
\item Well radius (m) rw <- 0.11
\item Casing radius (m): rc <- 0.11
\item Aquifer Thickness (m) B=400
}
}
\keyword{datasets}
