% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warren_root_utilities.R
\name{warren_root_calculate_parameters}
\alias{warren_root_calculate_parameters}
\title{warren_root_calculate_parameters}
\usage{
warren_root_calculate_parameters(ptest, par, hydraulic = TRUE)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{par}{A list with the parameters a, t0, t1, tm}

\item{hydraulic}{Logical flag to indicate if hydraulic parameters are calculated.
If False, the the statistcal parameter (a and t0) are calculated.}
}
\value{
A list with the parameters:
\itemize{
\item Tr: Transmissivity Tf
\item Ss: Storativity Sf
\item Sm: Storativity Sm
\item lambda: Interflow porosity
}
}
\description{
A function to calculates the hydraulic parameters for a Warren and Root model
}
\references{
Warren, J. & Root, P. The Behavior of Naturally Fractured Reservoirs. Society of
Petroleum Engineers Journal., 1963, 3.
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
