% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warren_root_utilities.R
\name{warren_root_solution}
\alias{warren_root_solution}
\title{warren_root_solution}
\usage{
warren_root_solution(ptest, a, t0, t1, tm, t)
}
\arguments{
\item{ptest}{A pumping_test object}

\item{a}{Slope of the Cooper-Jacob solution r}

\item{t0}{Intercept with the time axis for the early time asymptote}

\item{t1}{Intercept with the time axis for the late time asymptote}

\item{tm}{Time of the minimum derivative}

\item{t}{A numeric vector with the time values}
}
\value{
A numeric vector with the calculated drawdown
}
\description{
Function to calculate the drawdown using the Warren Root model
}
\references{
Warren, J. & Root, P. The Behavior of Naturally Fractured Reservoirs. Society of
Petroleum Engineers Journal., 1963, 3.
}
\seealso{
Other warren_root functions: \code{\link{warren_root_WF_LT}},
  \code{\link{warren_root_solution_dlogt}},
  \code{\link{warren_root_solution_initial}},
  \code{\link{warren_root_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{warren_root functions}
