% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warren_root_utilities.R
\name{warren_root_solution_initial}
\alias{warren_root_solution_initial}
\title{warren_root_solution_initial}
\usage{
warren_root_solution_initial(ptest)
}
\arguments{
\item{ptest}{A pumping_test object}
}
\value{
A list with
\itemize{
\item a: Slope
\item t0: Intercept of the early time asymptote
\item t1: Intercept of the late time asymptote
\item tm: time of the minimum derivative
}
}
\description{
Calculates the initial values of parameters of Warren and Root solution
}
\references{
Warren, J. & Root, P. The Behavior of Naturally Fractured Reservoirs. Society of
Petroleum Engineers Journal., 1963, 3.
}
\seealso{
Other warren_root functions: \code{\link{warren_root_WF_LT}},
  \code{\link{warren_root_solution_dlogt}},
  \code{\link{warren_root_solution}},
  \code{\link{warren_root_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{warren_root functions}
