% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warren_root_utilities.R
\name{warren_root_well_function}
\alias{warren_root_well_function}
\title{warren_root_well_function}
\usage{
warren_root_well_function(td, par)
}
\arguments{
\item{td}{A numeric vector with dimensionless time}

\item{par}{List with parameters sigma, interflow porosity (lambda) and Stehfest coefficients}
}
\value{
A numeric vector with the dimensionless drawdown
}
\description{
Calculates the well function of the Warrent and Root model
}
\examples{
td <- logseq(-1, 4, 50)
par <- list(sigma = 1, lambda = 1, coeffs = stehfest_coeffs(8))
W <- warren_root_well_function(td, par)
}
\references{
Warren, J. & Root, P. The Behavior of Naturally Fractured Reservoirs. Society of
Petroleum Engineers Journal., 1963, 3.
}
\seealso{
Other warren_root functions: \code{\link{warren_root_WF_LT}},
  \code{\link{warren_root_solution_dlogt}},
  \code{\link{warren_root_solution_initial}},
  \code{\link{warren_root_solution}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{warren_root functions}
