% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warren_root_utilities.R
\name{warren_root_WF_LT}
\alias{warren_root_WF_LT}
\title{warren_root_WF_LT}
\usage{
warren_root_WF_LT(p, arg1, arg2)
}
\arguments{
\item{p}{Laplace tranform parameter}

\item{arg1}{Value of parameter sigma}

\item{arg2}{Value of parameter lambda}
}
\value{
A numeric vector with the dimensionless drawdown in the Laplace domain
}
\description{
Calculates the well function of the Warrent and Root model
}
\examples{
p <- 0.5
sigma <- 1
lambda <- 1
s <- warren_root_WF_LT(p, sigma, lambda)
}
\references{
Warren, J. & Root, P. The Behavior of Naturally Fractured Reservoirs. Society of
Petroleum Engineers Journal., 1963, 3.
}
\seealso{
Other warren_root functions: \code{\link{warren_root_solution_dlogt}},
  \code{\link{warren_root_solution_initial}},
  \code{\link{warren_root_solution}},
  \code{\link{warren_root_well_function}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{warren_root functions}
