% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_dimension_utilities.R
\name{flow_dimension}
\alias{flow_dimension}
\title{flow_dimension}
\usage{
flow_dimension(t, s, d = 2, method = "central")
}
\arguments{
\item{t}{Numeric vector with time values}

\item{s}{Numeric vector with drawdown values}

\item{d}{Derivative parameter. if method equals to bourdet then d is equal to
the number of adjacent values used in the derivative calculation. If method is
equal to spline then d is equal to the number of knots used in the interpolation
of the drawdown data. In this case a value of d=20 to d=30 is recommended. If
method is equal to spane then d is equal to the number of points used in the linear
regression approach.}

\item{method}{Method to calculate the derivative. See log_derivative}
}
\value{
This function returns a list with components named as x and y that contains the
flow dimension y evaluated at specific points x.
}
\description{
Function to calculate the flow dimension using different approaches
}
\author{
Oscar Garcia-Cabrejo, \email{khaors@gmail.com}
}
\concept{flow_dimension functions}
